### 2020-02-13 #2096 @whedon
[PRE REVIEW]: Pippin: A pipeline for supernova cosmology
**Submitting author:** @samreay (<a href="http://orcid.org/0000-0003-2071-9349">Samuel Hinton</a>)
**Repository:** <a href="https://github.com/Samreay/Pippin" target ="_blank">https://github.com/Samreay/Pippin</a>
**Version:** 0.1.1

Thanks for submitting your paper to JOSS samreay. **Currently, there isn't an JOSS editor assigned** to your paper.

samreay if you have any suggestions for potential reviewers then please mention them here in this thread (without tagging them with an @). In addition, [this list of people](https://bit.ly/joss-reviewers) have already agreed to review for JOSS and may be suitable for this submission (please start at the bottom of the list).
