### 2020-02-13 #2096 @arfon
:wave: @Samreay - thanks for your submission to JOSS. Does this paper accompany a submission to a AAS journal?
