+++ configs/srh/srh_fakes_getphasecor.yml
@@ -1,0 +1,201 @@
+
+ANCHORS:  # The name of this doesnt actually matter, its all about the & and *
+
+  DEFAULT_SIM_MAP: &default_map
+    FLUXERRMODEL_FILE: $DES_ROOT/simlibs/maps/DES5YR_SIM_ERRORFUDGES_DIFFIMG_SBMAG_5.DAT
+    RANSEED_REPEAT: 1 12345
+    FLUXERRMODEL_REDCOV(DEEP): "g:0.1,r:0.1,i:0.1,z:0.1"
+    FLUXERRMODEL_REDCOV(SHALLOW): "griz:0.2"
+
+  FITOPTS: &default_fitopts
+    BATCH_INFO: sbatch $SBATCH_TEMPLATES/SBATCH_Midway2_10hr.TEMPLATE 10
+    APPEND_TABLE_TEXT: SNRMAX_g SNRMAX_r SNRMAX_i SNRMAX_z chi2_g chi2_r chi2_i chi2_z
+
+  SNLCINP: &default_mapdata
+    FLUXERRMODEL_FILE: $DES_ROOT/simlibs/maps/DES5YR_FAKES_ERRORFUDGES_DIFFIMG_SBMAG_5.DAT
+
+  DESLIKESIM: &deslikesim
+    IA_G10:
+      BASE: surveys/des/sims_ia/sn_ia_salt2_g10_des5yr.input
+    IAX:
+      BASE: surveys/des/sims_cc/sn_iax_dust.input
+    IA91BG:
+      BASE: surveys/des/sims_cc/sn_ia91bg.input
+    CC_TEMPLATES:
+      BASE: surveys/des/sims_cc/sn_collection_maria_dust1.input
+
+
+DATAPREP:
+  DATADES5YR:
+    OPTS:
+      RAW_DIR: $DES_ROOT/lcmerge/DESALL_forcePhoto_real_snana_fits
+      TYPES:
+        IA: [1, 101]
+        NONIA: [120, 130, 20, 30]
+
+
+SIM:
+
+  SIMDES5YR_NOPHASE:
+    <<: *deslikesim
+    GLOBAL:
+      <<: *default_map
+      NGEN_UNIT: 1
+      RANSEED_REPEAT: 5 12345
+
+  SIMDES5YR_02_0:
+    <<: *deslikesim
+    GLOBAL:
+      <<: *default_map
+      NGEN_UNIT: 1
+      RANSEED_REPEAT: 5 12345
+      GENMAG_SMEAR_ADDPHASECOR: 0.02 0
+
+  SIMDES5YR_02_1:
+    <<: *deslikesim
+    GLOBAL:
+      <<: *default_map
+      NGEN_UNIT: 1
+      RANSEED_REPEAT: 5 12345
+      GENMAG_SMEAR_ADDPHASECOR: 0.02 1
+
+  SIMDES5YR_02_3:
+    <<: *deslikesim
+    GLOBAL:
+      <<: *default_map
+      NGEN_UNIT: 1
+      RANSEED_REPEAT: 5 12345
+      GENMAG_SMEAR_ADDPHASECOR: 0.02 3
+
+  SIMDES5YR_02_6:
+    <<: *deslikesim
+    GLOBAL:
+      <<: *default_map
+      NGEN_UNIT: 1
+      RANSEED_REPEAT: 5 12345
+      GENMAG_SMEAR_ADDPHASECOR: 0.02 6
+
+
+
+  SIMDES5YR_03_0:
+    <<: *deslikesim
+    GLOBAL:
+      <<: *default_map
+      NGEN_UNIT: 1
+      RANSEED_REPEAT: 5 12345
[…]
