### 2020-02-14 #2096 @Samreay
Hi @arfon - not this time, JOSS only. The primary results from using Pippin will be published in about a years time for the DES 5YR analysis cosmology results. The initial plan was to wait until then to publish Pippin as a companion paper, but its had enough uptake in the community we realised we shouldn't wait.
