+++ configs/srh/srh_fakes_getphasecor.yml
@@ -12,8 +12,6 @@ ANCHORS:  # The name of this doesnt actually matter, its all about the & and *
   DESLIKESIM: &deslikesim
     IAX:
       BASE: surveys/des/sims_cc/sn_iax_dust.input
-    IA91BG:
-      BASE: surveys/des/sims_cc/sn_ia91bg.input
     CC_TEMPLATES:
       BASE: surveys/des/sims_cc/sn_collection_maria_dust1.input
     GLOBAL:
