+++ configs/srh/srh_fakes_getphasecor.yml
@@ -12,6 +12,8 @@ ANCHORS:  # The name of this doesnt actually matter, its all about the & and *
   DESLIKESIM: &deslikesim
     IAX:
       BASE: surveys/des/sims_cc/sn_iax_dust.input
+    IA91BG:
+      BASE: surveys/des/sims_cc/sn_ia91bg.input
     CC_TEMPLATES:
       BASE: surveys/des/sims_cc/sn_collection_maria_dust1.input
     GLOBAL:
@@ -38,14 +40,14 @@ SIM:
     <<: *deslikesim
     IA_G10:
       BASE: surveys/des/sims_ia/sn_ia_salt2_g10_des5yr.input
-      GENMAG_SMEAR_ADDPHASECOR: 0 0
+      GENMAG_SMEAR_ADDPHASECOR: 0 0.001
 
 
   SIMDES5YR_02_0:
     <<: *deslikesim
     IA_G10:
       BASE: surveys/des/sims_ia/sn_ia_salt2_g10_des5yr.input
-      GENMAG_SMEAR_ADDPHASECOR: 0.02 0
+      GENMAG_SMEAR_ADDPHASECOR: 0.02 0.001
 
   SIMDES5YR_02_1:
     <<: *deslikesim
@@ -78,7 +80,7 @@ SIM:
     <<: *deslikesim
     IA_G10:
       BASE: surveys/des/sims_ia/sn_ia_salt2_g10_des5yr.input
-      GENMAG_SMEAR_ADDPHASECOR: 0.03 0
+      GENMAG_SMEAR_ADDPHASECOR: 0.03 0.001
 
   SIMDES5YR_03_1:
     <<: *deslikesim
@@ -109,7 +111,7 @@ SIM:
     <<: *deslikesim
     IA_G10:
       BASE: surveys/des/sims_ia/sn_ia_salt2_g10_des5yr.input
-      GENMAG_SMEAR_ADDPHASECOR: 0.04 0
+      GENMAG_SMEAR_ADDPHASECOR: 0.04 0.001
 
   SIMDES5YR_04_1:
     <<: *deslikesim
