### 2020-02-14 96af7ed
Adding travis config
