+++ .travis.yml
@@ -1,0 +1,20 @@
+language: python
+
+python:
+  - "3.7"
+  - "3.8"
+
+before_install:
+  - python -m pip install --upgrade pip setuptools wheel
+  - pip install -r requirements.txt
+  - pip install -r requirements_dev.txt
+install:
+  - python setup.py sdist
+script:
+  - pytest -v .
+
+notifications:
+  email:
+    recipients:
+      - samuelreay@gmail.com
+    on_failure: always
