+++ .travis.yml
@@ -8,6 +8,7 @@ before_install:
   - python -m pip install --upgrade pip setuptools wheel
   - pip install -r requirements.txt
   - pip install -r requirements_dev.txt
+  - pip install -U numpy
 install:
   - python setup.py sdist
 script:
