### 2020-02-14 985579e
Moving groupid into the except block
