### 2020-02-14 ab160c2
Moving groupinfo again
