### 2020-02-14 c04da55
Hiding vaild config test in readme until clicked on
