### 2020-02-14 043376e
Testing 0.05 scatter as well
