+++ configs/srh/srh_fakes_getphasecor.yml
@@ -13,6 +13,7 @@ ANCHORS:  # The name of this doesnt actually matter, its all about the & and *
       RANSEED_REPEAT: 5 12345
       FLUXERRMODEL_REDCOV(DEEP): "g:0.1,r:0.1,i:0.1,z:0.1"
       FLUXERRMODEL_REDCOV(SHALLOW): "griz:0.2"
+      SEARCHEFF_zHOST_FILE: $DES_ROOT/models/searcheff/samtest/efficiency_zHD_G10.dat
       NGEN_UNIT: 1
 
 
