### 2020-02-14 #2096 @arfon
@temuller @kbarbary - would you be willing to review this submission for JOSS?
