### 2020-02-14 #2096 @arfon
Thanks for your interest @temuller. The software appears to be mostly written in Python 3.7 (https://github.com/Samreay/Pippin).

> How much time do I have to review?

Ideally we'd like to have the review complete in 2-3 weeks. The review is checklist-driven and happens on GitHub. We have pretty detailed guidelines [here](https://joss.readthedocs.io/en/latest/reviewer_guidelines.htmll) for reviewers.
