### 2020-02-14 b55196c
Adding obs data to two of the sims
