### 2020-02-15 #2096 @kbarbary
Hey Arfon! Thanks for the invite. Unfortunately  I won't have the time to give this a thorough review. (I'm now a couple years removed from SN cosmology work.)
