### 2020-02-17 d6fe553
Removing private data path from clumping file
