### 2020-02-17 bce6708
Dirty hack before I ask Rick to fix double defining paths
