+++ pippin/dataprep.py
@@ -133,7 +133,7 @@ fi
 
         command_string = self.clump_command.format(genversion=self.genversion, data_path=self.data_path)
 
-        if self.data_path == "/scratch/midway2/rkessler/SNDATA_ROOT/SIM":
+        if "lcmerge" not in self.data_path:
             command_string.replace("PRIVATE_DATA_PATH", "! PRIVATE_DATA_PATH")
 
         format_dict = {"job_name": self.job_name, "log_file": self.logfile, "path_to_task": self.path_to_task, "done_file": self.done_file}
