### 2020-02-17 3f56406
Will add the simulation readme check after a few weeks so I dont break everyone's sims.
