### 2020-02-17 f525872
Testing revised search eff map
