+++ configs/srh/srh_compute_zhost.yml
@@ -27,6 +27,7 @@ SIM:
       APPLY_SEARCHEFF_OPT: 1
       NGEN_UNIT: 0.4
       RANSEED_REPEAT: 25 12345
+      SEARCHEFF_PIPELINE_EFF_FILE: $SNANA_DEBUG/SIM_PHOTPROB/SEARCHEFF_PIPELINE_DES_REVISED.DAT
 
   DESSIM_DEEPX1:
     <<: *des
@@ -36,7 +37,7 @@ SIM:
       APPLY_SEARCHEFF_OPT: 1
       NGEN_UNIT: 0.4
       RANSEED_REPEAT: 25 12345
-
+      SEARCHEFF_PIPELINE_EFF_FILE: $SNANA_DEBUG/SIM_PHOTPROB/SEARCHEFF_PIPELINE_DES_REVISED.DAT
 LCFIT:
   DSPEC:
     BASE: surveys/des/lcfit_nml/des_5yr.nml
