### 2020-02-18 cd86c20
Adding sim without photprob
