+++ configs/srh/srh_compute_zhost.yml
@@ -29,6 +29,15 @@ SIM:
       RANSEED_REPEAT: 25 12345
       SEARCHEFF_PIPELINE_EFF_FILE: $SNANA_DEBUG/SIM_PHOTPROB/SEARCHEFF_PIPELINE_DES_REVISED.DAT
 
+  DESSIM_SVA_NOPHOTPROB:
+    <<: *des
+    GLOBAL:
+      SEARCHEFF_zHOST_FILE: NONE
+      APPLY_SEARCHEFF_OPT: 1
+      NGEN_UNIT: 0.4
+      RANSEED_REPEAT: 25 12345
+      SEARCHEFF_PIPELINE_EFF_FILE: $SNANA_DEBUG/SIM_PHOTPROB/SEARCHEFF_PIPELINE_DES_REVISED.DAT
+
   DESSIM_DEEPX1:
     <<: *des
     GLOBAL:
