### 2020-02-18 e0ca392
Removing search eff
