### 2020-02-18 7e8102c
Recomputing zhost after adding CUTVAL to the map
