+++ configs/srh/srh_compute_zhost.yml
@@ -29,23 +29,15 @@ SIM:
       RANSEED_REPEAT: 25 12345
       SEARCHEFF_PIPELINE_EFF_FILE: $SNANA_DEBUG/SIM_PHOTPROB/SEARCHEFF_PIPELINE_DES_REVISED.DAT
 
-  DESSIM_SVA_NOPHOTPROB:
-    <<: *des
-    GLOBAL:
-      SEARCHEFF_zHOST_FILE: NONE
-      APPLY_SEARCHEFF_OPT: 1
-      NGEN_UNIT: 0.4
-      RANSEED_REPEAT: 25 12345
-
-  DESSIM_DEEPX1:
-    <<: *des
-    GLOBAL:
-      SEARCHEFF_zHOST_FILE: NONE
-      HOSTLIB_FILE:    $DES_ROOT/simlibs/DES_5YR.HOSTLIB
-      APPLY_SEARCHEFF_OPT: 1
-      NGEN_UNIT: 0.4
-      RANSEED_REPEAT: 25 12345
-      SEARCHEFF_PIPELINE_EFF_FILE: $SNANA_DEBUG/SIM_PHOTPROB/SEARCHEFF_PIPELINE_DES_REVISED.DAT
+#  DESSIM_DEEPX1:
+#    <<: *des
+#    GLOBAL:
+#      SEARCHEFF_zHOST_FILE: NONE
+#      HOSTLIB_FILE:    $DES_ROOT/simlibs/DES_5YR.HOSTLIB
+#      APPLY_SEARCHEFF_OPT: 1
+#      NGEN_UNIT: 0.4
+#      RANSEED_REPEAT: 25 12345
+#      SEARCHEFF_PIPELINE_EFF_FILE: $SNANA_DEBUG/SIM_PHOTPROB/SEARCHEFF_PIPELINE_DES_REVISED.DAT
 LCFIT:
   DSPEC:
     BASE: surveys/des/lcfit_nml/des_5yr.nml
@@ -56,28 +48,28 @@ LCFIT:
       BATCH_INFO: sbatch $SBATCH_TEMPLATES/SBATCH_Midway2_10hr.TEMPLATE 50
       APPEND_TABLE_TEXT: HOST_MAG_i HOST_MAG_r HOST_MAG_g
 
-  DPHOTO:
-    BASE: surveys/des/lcfit_nml/des_5yr.nml
-    MASK: DES
-    SNLCINP:
-      ROOTFILE_OUT: FIT.ROOT
-    FITINP:
-      DOFIT_PHOTOZ: T
-      OPT_PHOTOZ: 1
-      PRIOR_ZERRSCALE: 1.0E6
-      PHOTOZ_BOUND:  0.00, 1.5
-    OPTS:
-      BATCH_INFO: sbatch $SBATCH_TEMPLATES/SBATCH_Midway2_10hr.TEMPLATE 50
-      APPEND_TABLE_TEXT: HOST_MAG_i HOST_MAG_r HOST_MAG_g
+#  DPHOTO:
+#    BASE: surveys/des/lcfit_nml/des_5yr.nml
+#    MASK: DES
+#    SNLCINP:
+#      ROOTFILE_OUT: FIT.ROOT
+#    FITINP:
+#      DOFIT_PHOTOZ: T
+#      OPT_PHOTOZ: 1
+#      PRIOR_ZERRSCALE: 1.0E6
+#      PHOTOZ_BOUND:  0.00, 1.5
+#    OPTS:
[…]
