### 2020-02-18 41ee336
Creating sim comparisons for DES and LOWZ
