### 2020-02-18 ef5bd86
Creating sim comparisons for DES and LOWZ
