+++ configs/srh/srh_final_sim_dist_check.yml
@@ -62,6 +62,8 @@ LCFIT:
   L:
     BASE: surveys/lowz/lcfit_nml/lowz.nml
     MASK: LOWZ
+    SNLCINP:
+      ROOTFILE_OUT: FIT.ROOT
     OPTS:
       BATCH_INFO: sbatch $SBATCH_TEMPLATES/SBATCH_Midway2_10hr.TEMPLATE 20
       APPEND_TABLE_TEXT: SNRMAX_g SNRMAX_r SNRMAX_i SNRMAX_z chi2_g chi2_r chi2_i chi2_z
@@ -70,6 +72,8 @@ LCFIT:
   D:
     BASE: surveys/des/lcfit_nml/des_5yr.nml
     MASK: DES
+    SNLCINP:
+      ROOTFILE_OUT: FIT.ROOT
     OPTS:
       BATCH_INFO: sbatch $SBATCH_TEMPLATES/SBATCH_Midway2_10hr.TEMPLATE 20
       APPEND_TABLE_TEXT: SNRMAX_g SNRMAX_r SNRMAX_i SNRMAX_z chi2_g chi2_r chi2_i chi2_z
