+++ configs/srh/srh_final_sim_dist_check.yml
@@ -66,7 +66,7 @@ LCFIT:
       ROOTFILE_OUT: FIT.ROOT
     OPTS:
       BATCH_INFO: sbatch $SBATCH_TEMPLATES/SBATCH_Midway2_10hr.TEMPLATE 20
-      APPEND_TABLE_TEXT: SNRMAX_g SNRMAX_r SNRMAX_i SNRMAX_z chi2_g chi2_r chi2_i chi2_z
+      APPEND_TABLE_TEXT: SNRMAX_g SNRMAX_r SNRMAX_i SNRMAX_z
 
 
   D:
