+++ configs/srh/srh_final_sim_dist_check.yml
@@ -62,11 +62,8 @@ LCFIT:
   L:
     BASE: surveys/lowz/lcfit_nml/lowz.nml
     MASK: LOWZ
-    SNLCINP:
-      ROOTFILE_OUT: FIT.ROOT
     OPTS:
       BATCH_INFO: sbatch $SBATCH_TEMPLATES/SBATCH_Midway2_10hr.TEMPLATE 20
-      APPEND_TABLE_TEXT: SNRMAX_g SNRMAX_r SNRMAX_i SNRMAX_z
 
 
   D:
