### 2020-02-18 6b12857
Updating plot_histograms
