### 2020-02-19 4d221a0
Reducing batch size to 20 and rerunning
