### 2020-02-20 9d50ca2
Updating covariance
