+++ configs/srh/srh_fakes_checkmaps.yml
@@ -16,7 +16,7 @@ ANCHORS:  # The name of this doesnt actually matter, its all about the & and *
     RANSEED_REPEAT: 1 12345
 
   DEFAULT_CORR: &default_corr
-    FLUXERRMODEL_REDCOV(DEEP): "g:0.1,r:0.1,i:0.1,z:0.1"
+    FLUXERRMODEL_REDCOV(DEEP): "g:0.05,r:0.05,i:0.05,z:0.05"
     FLUXERRMODEL_REDCOV(SHALLOW): "griz:0.2"
 
   FAKES_INPUT: &fakesinput
