### 2020-02-20 #2096 @arfon
:wave: @daniel-muthukrishna - would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html
