### 2020-02-20 4f0e2e2
Updating checkmaps to use different partition
