### 2020-02-20 2b3fbf6
Updating checkmaps to use different partition
