+++ configs/srh/srh_fakes_checkmaps.yml
@@ -12,7 +12,7 @@
 ANCHORS:  # The name of this doesnt actually matter, its all about the & and *
 
   DEFAULT_SIM_MAP: &default_map
-    FLUXERRMODEL_FILE: $DES_ROOT/simlibs/maps/DES5YR_SIM_ERRORFUDGES_DIFFIMG_LOGSNR_SBMAG_5.DAT
+    FLUXERRMODEL_FILE: $DES_ROOT/simlibs/maps/DES5YR_SIM_ERRORFUDGES_DIFFIMG_SBMAG_5.DAT
     RANSEED_REPEAT: 1 12345
 
   DEFAULT_CORR: &default_corr
@@ -27,7 +27,7 @@ ANCHORS:  # The name of this doesnt actually matter, its all about the & and *
     APPEND_TABLE_TEXT: SNRMAX_g SNRMAX_r SNRMAX_i SNRMAX_z chi2_g chi2_r chi2_i chi2_z
 
   SNLCINP: &default_mapdata
-    FLUXERRMODEL_FILE: $DES_ROOT/simlibs/maps/DES5YR_FAKES_ERRORFUDGES_DIFFIMG_LOGSNR_SBMAG_5.DAT
+    FLUXERRMODEL_FILE: $DES_ROOT/simlibs/maps/DES5YR_FAKES_ERRORFUDGES_DIFFIMG_SBMAG_5.DAT
 
   DESLIKESIM: &deslikesim
     IA_G10:
@@ -71,7 +71,7 @@ SIM:
 
   # Great, so that should allow a good mix for the correlation to be determined
   # broken down by deep and shallow fields. Lets simulate this mix
-  SIMFAKE_2D_MIX:
+  SIMFAKE_1D_MIX:
     IA: *fakesinput
     GLOBAL:
       <<: *default_map
@@ -171,7 +171,7 @@ LCFIT:
       FLUXERRMODEL_FILE: NONE
     OPTS: *default_fitopts
 
-  F2D: # fit fakes using the 2D correction, which only apples to the real fakes, not the sims
+  F1D: # fit fakes using the 2D correction, which only apples to the real fakes, not the sims
     BASE: surveys/des/lcfit_nml/fakes.nml
     MASK: FAKE
     SNLCINP: *default_mapdata
@@ -215,22 +215,15 @@ ANALYSE:
 
   CHECK_FAKES:
     HISTOGRAM:
-      - F2D_FAKES
-      - F2D_SIMFAKE_2D_MIX
+      - F1D_FAKES
+      - F1D_SIMFAKE_1D_MIX
 
   CHECK_FAKES_BEFORE_AFTER:
     HISTOGRAM:
       - F0D_FAKES
-      - F2D_SIMFAKES_NOCORR
-      - F2D_FAKES
-      - F2D_SIMFAKE_2D_MIX
-
-  CHECK_3YR_DIST:
-    HISTOGRAM:
-      - D2D_DATADES3YR
-      - D2D_SIMDES3YR_2D_REDCOV
-      - D2D_SIMDES3YR_2D_INDCOV
-      - D2D_SIMDES3YR_2D_NOCOR
+      - F1D_SIMFAKES_NOCORR
+      - F1D_FAKES
+      - F1D_SIMFAKE_1D_MIX
 
   CHECK_5YR_DIST_1D:
     HISTOGRAM:
