### 2020-02-21 091af88
Updating plot_histogram code to remove fitprob<0.05 always
