### 2020-02-21 2164e5b
Updating check_fakes
