+++ configs/srh/srh_fakes_checkmaps.yml
@@ -17,7 +17,7 @@ ANCHORS:  # The name of this doesnt actually matter, its all about the & and *
 
   DEFAULT_CORR: &default_corr
     FLUXERRMODEL_REDCOV(DEEP): "g:0.05,r:0.05,i:0.05,z:0.05"
-    FLUXERRMODEL_REDCOV(SHALLOW): "griz:0.2"
+    FLUXERRMODEL_REDCOV(SHALLOW): "griz:0.25"
 
   FAKES_INPUT: &fakesinput
     BASE: surveys/des/sims_ia/fakes_diffimg.input
