### 2020-02-21 3122c4d
No longer putting fitprob on a log scale
