### 2020-02-21 dfe18b5
Halving size of GLOBAL_ML training sim
