+++ configs/global_ml.yml
@@ -6,7 +6,7 @@ SIM:
       BASE: surveys/des/sims_cc/sn_collection_maria_split_a.input
     GLOBAL:
       NGEN_UNIT: 5
-      RANSEED_REPEAT: 40 12345
+      RANSEED_REPEAT: 20 12345
 
   TESTSIM:
     IA_C11:
@@ -14,7 +14,7 @@ SIM:
     II:
       BASE: surveys/des/sims_cc/sn_collection_maria_split_b.input
     GLOBAL:
-      NGEN_UNIT: 3
+      NGEN_UNIT: 1
       RANSEED_REPEAT: 20 12345
 
 LCFIT:
