### 2020-02-24 2bb7dd1
Adding prior variant to cosmomc
