+++ data_files/cosmomc_templates/sn_omw_prior.ini
@@ -1,0 +1,20 @@
+#DEFAULT({path_to_cosmomc}/batch2/BAO.ini)
+DEFAULT({path_to_cosmomc}/batch2/JLA.ini)
+#DEFAULT({path_to_cosmomc}/batch2/GAUSS.ini)
+#DEFAULT({path_to_cosmomc}/batch2/common.ini)
+DEFAULT({path_to_cosmomc}/batch1/common_batch1_SNalone.ini)
+INCLUDE({ini_dir}/base.ini)
+
+parameterization=background
+
+MPI_Converge_Stop = 0.002
+MPI_Limit_Converge = 0.002
+MPI_Limit_Converge_Err = 0.002
+
+param[wa]=0
+param[omegak]=0
+param[w]=-0.995 -3 1.0 0.02 0.02
+param[omegam]=0.3 0.0 1.0 0.002 0.002
+action = 0
+
+prior[omegam] = 0.3 .01
