### 2020-02-24 6a79852
Adding prior variant to cosmomc
