+++ configs/srh/srh_des5yr_errorbudget.yml
@@ -1,12 +1,4 @@
 DATAPREP:
-  DATADES5YR:  # 5 yr photometry sample
-    OPTS:
-      RAW_DIR: $DES_ROOT/lcmerge/DESALL_forcePhoto_real_snana_fits
-      TYPES:
-        IA: [1, 101]
-        NONIA: [120, 130, 20, 30]
-      BLIND: True
-
   DESSIMBIAS5YRIA_C11:
     OPTS:
       RAW_DIR: $SCRATCH_SIMDIR/PIP_GLOBAL_BIASCOR_DESSIMBIAS5YRIA_C11
@@ -16,6 +8,20 @@ DATAPREP:
   DESSIMBIAS5YRCC:
     OPTS:
       RAW_DIR: $SCRATCH_SIMDIR/PIP_GLOBAL_BIASCOR_DESSIMBIAS5YRCC
+
+SIM:
+  DATADES:
+    IA_G10:
+      BASE: surveys/des/sims_ia/sn_ia_salt2_g10_des5yr.input
+    IAX:
+      BASE: surveys/des/sims_cc/sn_iax_dust.input
+    IA91BG:
+      BASE: surveys/des/sims_cc/sn_ia91bg.input
+    CC_TEMPLATES:
+      BASE: surveys/des/sims_cc/sn_collection_maria_dust1.input
+    GLOBAL:
+      NGEN_UNIT: 0.1
+      RANSEED_REPEAT: 10 12345
 
 LCFIT:
   DS:
@@ -61,13 +67,6 @@ BIASCOR:
     SIMFILE_BIASCOR: [D_DESSIMBIAS5YRIA_G10]
     SIMFILE_CCPRIOR: D_DESSIMBIAS5YRCC
     CLASSIFIER: SNNTEST
-    MUOPTS:
-      SCATTER_C11:
-        SIMFILE_BIASCOR: [D_DESSIMBIAS5YRIA_C11]
-        FITOPT: 0
-      CLAS_SNIRF:
-        CLASSIFIER: SNIRFTEST
-        FITOPT: 0
     OPTS:
       BATCH_INFO: sbatch $SBATCH_TEMPLATES/SBATCH_Midway2_10hr.TEMPLATE 5
 
@@ -76,17 +75,18 @@ CREATE_COV:
     OPTS:
       COVOPTS:
         - "[NOSYS] [=DEFAULT,=DEFAULT]"
-        - "[SCATTER] [=DEFAULT,+SCATTER]"
-        - "[CLASSIFICATION] [=DEFAULT,+CLAS]"
 
 
 COSMOMC:
   SN_OMW:
     OPTS:
       INI: sn_omw_prior
+      COVOPTS: [ALL, NOSYS]
 
 
 ANALYSE:
   ALL_OMW:
+    MASK_COSMOMC: OMW
     OPTS:
-
+      BLIND: w
+      SHIFT: True
