+++ configs/srh/srh_des5yr_errorbudget.yml
@@ -10,7 +10,7 @@ DATAPREP:
       RAW_DIR: $SCRATCH_SIMDIR/PIP_GLOBAL_BIASCOR_DESSIMBIAS5YRCC
 
 SIM:
-  DATADES:
+  DATADES5YR:
     IA_G10:
       BASE: surveys/des/sims_ia/sn_ia_salt2_g10_des5yr.input
     IAX:
@@ -67,6 +67,13 @@ BIASCOR:
     SIMFILE_BIASCOR: [D_DESSIMBIAS5YRIA_G10]
     SIMFILE_CCPRIOR: D_DESSIMBIAS5YRCC
     CLASSIFIER: SNNTEST
+    MUOPTS:
+      SCATTER_C11:
+        SIMFILE_BIASCOR: [D_DESSIMBIAS5YRIA_C11]
+        FITOPT: 0
+      CLAS_SNIRF:
+        CLASSIFIER: SNIRFTEST
+        FITOPT: 0
     OPTS:
       BATCH_INFO: sbatch $SBATCH_TEMPLATES/SBATCH_Midway2_10hr.TEMPLATE 5
 
@@ -75,18 +82,17 @@ CREATE_COV:
     OPTS:
       COVOPTS:
         - "[NOSYS] [=DEFAULT,=DEFAULT]"
+        - "[SCATTER] [=DEFAULT,+SCATTER]"
+        - "[CLASSIFICATION] [=DEFAULT,+CLAS]"
 
 
 COSMOMC:
   SN_OMW:
     OPTS:
       INI: sn_omw_prior
-      COVOPTS: [ALL, NOSYS]
 
 
 ANALYSE:
   ALL_OMW:
-    MASK_COSMOMC: OMW
     OPTS:
-      BLIND: w
-      SHIFT: True
+
