+++ configs/srh/srh_des5yr_errorbudget.yml
@@ -63,7 +63,7 @@ MERGE:
 
 BIASCOR:
   5YR:
-    DATA: [DS_DATADES5YR]
+    DATA: [DS_SIMDATADES5YR]
     SIMFILE_BIASCOR: [D_DESSIMBIAS5YRIA_G10]
     SIMFILE_CCPRIOR: D_DESSIMBIAS5YRCC
     CLASSIFIER: SNNTEST
