+++ configs/srh/srh_des5yr_errorbudget.yml
@@ -94,5 +94,5 @@ COSMOMC:
 
 ANALYSE:
   ALL_OMW:
-    OPTS:
+    MASK_COSMOMC: OMW
 
