### 2020-02-24 8778d12
Allowing for arbitrary plotting scripts, and removing the donefile requirement
