### 2020-02-24 376ae66
Adidng chi2 text to plot output when plotting histograms
