### 2020-02-24 af9c493
Testing chi2 values on histogram plots
