+++ configs/srh/srh_des5yr_errorbudget.yml
@@ -90,6 +90,7 @@ COSMOMC:
   SN_OMW:
     OPTS:
       INI: sn_prior_omw
+      WALKERS: 20
 
 
 ANALYSE:
