+++ data_files/cosmomc_templates/sn_prior_omw.ini
@@ -7,9 +7,9 @@ INCLUDE({ini_dir}/base.ini)
 
 parameterization=background
 
-MPI_Converge_Stop = 0.002
-MPI_Limit_Converge = 0.002
-MPI_Limit_Converge_Err = 0.002
+MPI_Converge_Stop = 0.001
+MPI_Limit_Converge = 0.001
+MPI_Limit_Converge_Err = 0.001
 
 param[wa]=0
 param[omegak]=0
