### 2020-02-24 653cb3f
Adding error budget code whenever cosmomc is found
