### 2020-02-24 36d7f95
Adding num_walkers
