### 2020-02-24 fa20b0e
A whole lot of debug text coming right up
