### 2020-02-24 6b01a09
Fixing num jobs computation
