+++ pippin/manager.py
@@ -109,20 +109,16 @@ class Manager:
     def get_task_to_run(self, tasks_to_run, done_tasks):
         for t in tasks_to_run:
             can_run = True
-            if isinstance(t, CosmoMC):
-                print("DDDD ", t.dependencies, [d in done_tasks for d in t.dependencies])
             for dep in t.dependencies:
                 if dep not in done_tasks:
                     can_run = False
-                    print("ZZZZZZZZZZZ", dep)
-            print("EEEEEE ", can_run)
             if t.gpu and self.num_jobs_queue_gpu + t.num_jobs >= self.max_jobs_in_queue_gpu:
+                self.logger.warning(f"Cant submit {t} because num jobs {t.num_jobs} would take us over the limit with {self.num_jobs_queue_gpu} already running")
                 can_run = False
-            print("FFFFFF ", can_run)
 
             if not t.gpu and self.num_jobs_queue + t.num_jobs >= self.max_jobs_in_queue:
+                self.logger.warning(f"Cant submit {t} because num jobs {t.num_jobs} would take us over the limit with {self.num_jobs_queue} already running")
                 can_run = False
-            print("GGGGGG ", can_run)
 
             if can_run:
                 return t
@@ -201,7 +197,6 @@ class Manager:
             for t in running_tasks:
                 try:
                     completed = self.check_task_completion(t, blocked_tasks, done_tasks, failed_tasks, running_tasks, squeue)
-                    print(f"AAAAAAAAAAAAA Task {t} completed")
                     small_wait = small_wait or completed
                 except Exception as e:
                     self.logger.exception(e, exc_info=True)
@@ -209,10 +204,8 @@ class Manager:
 
             # Submit new jobs if needed
             while self.num_jobs_queue < self.max_jobs:
-                print(f"BBBBBBBBB Looking for more jobs please")
 
                 t = self.get_task_to_run(self.tasks, done_tasks)
-                print(f"CCCCCCCC Got task to run {t}")
                 if t is not None:
                     self.logger.info("")
                     self.tasks.remove(t)
