### 2020-02-24 6d085e8
Updating location of text in the histograms
