### 2020-02-24 b755752
Lets run with lowz as well
