### 2020-02-25 4955ba8
Updating ERRSCALE to be all zero
