+++ configs/srh/srh_deslowz_errorbudget.yml
@@ -99,6 +99,7 @@ BIASCOR:
     SIMFILE_BIASCOR: [D_DESSIMBIAS5YRIA_G10, L_LOWZSIMBIAS_G10]
     SIMFILE_CCPRIOR: D_DESSIMBIAS5YRCC
     CLASSIFIER: SNNTEST
+    MAKE_ALL_HUBBLE: True
     MUOPTS:
       SCATTER_C11:
         SIMFILE_BIASCOR: [D_DESSIMBIAS5YRIA_C11, L_LOWZSIMBIAS_C11]
@@ -159,4 +160,5 @@ COSMOMC:
 ANALYSE:
   ALL_OMW:
     MASK_COSMOMC: OMW
+    HISTOGRAM: [DS_SIMDATADES5YR, D_DESSIMBIAS5YRIA_G10, D_DESSIMBIAS5YRCC, D_DESSIMBIAS5YRIA_C11]
 
