+++ README.md
@@ -518,7 +518,7 @@ systematics, and determining which covariance options to run with.
 CREATE_COV:
   SOMELABEL:
     OPTS:
-      SYS_SCALE: create_cov/sys_scale.LIST  # Location of systematic scaling file
+      SYS_SCALE: surveys/des/bbc/scale_5yr.list  # Location of systematic scaling file
       FITOPT_SCALES:  # Optional
         FITOPT_LABEL: some_scale  # Note this is a partial match, ie SALT2: 1.0 would apply to all SALT2 cal fitopts
        MUOPT_SCALES:
