+++ examples/createcov.yml
@@ -7,7 +7,7 @@ CREATE_COV:
 
   SYSTEMATICS:
     OPTS:
-      SYS_SCALE: $somewhere/sys_scale.LIST # Location of sys_scale.LIST file, relative to create_cov dir by default
+      SYS_SCALE: $somewhere/sys_scale.list # Location of sys_scale.LIST file, relative to create_cov dir by default
       FITOPT_SCALES:  # Optional
         cal: 2.0  # checks if its in the label, not an exact match. So this would scale all *cal* fitopts by a factor of two
       MUOPT_SCALES: # Optional - used to construct SYSFILE input
