+++ configs/srh/srh_deslowz_errorbudget.yml
@@ -28,8 +28,8 @@ SIM:
     GLOBAL:
       NGEN_UNIT: 0.1
       RANSEED_REPEAT: 10 12345
-      HOSTLIB_FILE: DES-NEWMASS-v3.HOSTLIB
-      HOSTLIB_WGTMAP_FILE: OG_WGTMAP.DAT
+      HOSTLIB_FILE: $DES_ROOT/simlibs/DES-NEWMASS-v3.HOSTLIB
+      HOSTLIB_WGTMAP_FILE: $DES_ROOT/simlibs/OG_WGTMAP.DAT
       HOSTLIB_MSKOPT: 326
 
   SIMDATALOWZ:
