### 2020-02-25 44712bc
Adding zcut test to look at SALT2 systematics
