### 2020-02-25 3cfbddf
Adding zcut test to look at SALT2 systematics
