### 2020-02-25 931843b
Adding training test set into the mix
