+++ configs/srh/srh_training_sets.yml
@@ -45,6 +45,19 @@ SIM:
       NGEN_UNIT: 1
       RANSEED_REPEAT: 10 12345
 
+  G10_LI_GAUSS:
+    IA:
+      BASE: surveys/des/sims_ia/sn_ia_salt2_g10_des5yr.input
+    II_TEMPLATES:
+      BASE: surveys/des/sims_cc/sn_ii_maria_li_revised.input
+      DNDZ: 0.7*CC_S15
+    Ibc_TEMPLATES:
+      BASE: surveys/des/sims_cc/sn_ibc_maria_li_revised.input
+      DNDZ: 0.3*CC_S15
+    GLOBAL:
+      NGEN_UNIT: 1
+      RANSEED_REPEAT: 10 12345
+
 LCFIT:
   TESTFIT:
     BASE: surveys/des/lcfit_nml/des_5yr.nml
