### 2020-02-25 b71bb78
Removing mask
