### 2020-02-25 f82d95a
Checking other maps for zhost
