### 2020-02-25 581a895
Fixing the APPEND_TABLE_TEXT
