+++ configs/srh/srh_check_zhost.yml
@@ -50,7 +50,7 @@ LCFIT:
       ROOTFILE_OUT: FIT.ROOT
     OPTS:
       BATCH_INFO: sbatch $SBATCH_TEMPLATES/SBATCH_Midway2_10hr.TEMPLATE 50
-      APPEND_FITRES: MWEBV HOST_MAG_i HOST_MAG_r HOST_MAG_g SNRMAX_g SNRMAX_r SNRMAX_i SNRMAX_z
+      APPEND_TABLE_TEXT: HOST_MAG_i HOST_MAG_r HOST_MAG_g SNRMAX_g SNRMAX_r SNRMAX_i SNRMAX_z
 
 ANALYSE:
   HIST_COMPARISON_ALL:
