### 2020-02-25 c401de2
Adding consistent way of scanning log files
