### 2020-02-25 cbf371d
Separating classifier aggregation
