### 2020-02-25 8f8e35c
Reducing the size of the SALT2 calibration systematics
