### 2020-02-25 17a6b8f
Reducing the size of the SALT2 calibration systematics
