### 2020-02-25 445e68f
Reducing the size of the SALT2 calibration systematics
