+++ configs/srh/srh_deslowz_errorbudget.yml
@@ -97,16 +97,6 @@ MERGE:
   MERGE:
 
 BIASCOR:
-  ZCUT:
-    DATA: [DS_SIMDATADES5YR, LS_SIMDATALOWZ]
-    SIMFILE_BIASCOR: [D_DESSIMBIAS5YRIA_G10, L_LOWZSIMBIAS_G10]
-    SIMFILE_CCPRIOR: D_DESSIMBIAS5YRCC
-    CLASSIFIER: SNNTEST
-    MAKE_ALL_HUBBLE: True
-    OPT:
-      zmax: 0.75
-      BATCH_INFO: sbatch $SBATCH_TEMPLATES/SBATCH_Midway2_10hr.TEMPLATE 20
-
   5YR:
     DATA: [DS_SIMDATADES5YR, LS_SIMDATALOWZ]
     SIMFILE_BIASCOR: [D_DESSIMBIAS5YRIA_G10, L_LOWZSIMBIAS_G10]
@@ -172,13 +162,6 @@ CREATE_COV:
         - "[CHI2MAX_100] [=DEFAULT,=CHI2MAX_100]"
         - "[CCPRIORBINS] [=DEFAULT,=CCPRIORBINS]"
         - "[HOSTMASS_DBETA] [=DEFAULT,=HOSTMASS_DBETA]"
-
-  ZCUT:
-    MASK: ZCUT
-    OPTS:
-      COVOPTS:
-        - "[NOSYS] [=DEFAULT,=DEFAULT]"
-        - "[SALT] [+SALT,=DEFAULT]"
 
 COSMOMC:
   SN_OMW:
