### 2020-02-25 f31db6b
Reducing the size of the SALT2 calibration systematics
