### 2020-02-26 9254a6c
Adding better debug output
