### 2020-02-26 8fd55bd
Turning off scatter for sims and checking what happens
