### 2020-02-26 814a7a4
Updating default normalisation for SNN.
