### 2020-02-26 2e49798
Putting redshift flag in the training step instead of data prep
