### 2020-02-26 b345a54
Putting norm flag in the training step instead of data prep
