### 2020-02-26 f8e386f
Putting norm flag in the training step instead of data prep
