### 2020-02-26 ce01d64
Updating analyse, biascor and plotting scripts
