+++ pippin/analyse.py
@@ -61,7 +61,8 @@ class AnalyseChains(Task):  # TODO: Define the location of the output so we can 
         if isinstance(self.covopts, str):
             self.covopts = [self.covopts]
 
-        self.files = []
+        self.cosmomc_input_files = []
+        self.cosmomc_output_files = []
         self.names = []
         self.params = []
 
@@ -76,11 +77,14 @@ class AnalyseChains(Task):  # TODO: Define the location of the output so we can 
         self.lcfit_deps = self.get_deps(SNANALightCurveFit)
 
         if self.cosmomc_deps:
+            self.add_plot_script_to_run("parse_cosmomc.py")
             self.add_plot_script_to_run("plot_cosmomc.py")
-            self.add_plot_script_to_run("errbudget.py")
+            self.add_plot_script_to_run("plot_errbudget.py")
         if self.biascor_deps:
+            self.add_plot_script_to_run("parse_biascor.py")
             self.add_plot_script_to_run("plot_biascor.py")
         if self.lcfit_deps:
+            self.add_plot_script_to_run("parse_lcfit.py")
             self.add_plot_script_to_run("plot_histogram.py")
             self.add_plot_script_to_run("plot_efficiency.py")
 
@@ -94,15 +98,46 @@ class AnalyseChains(Task):  # TODO: Define the location of the output so we can 
         for c in self.cosmomc_deps:
             for covopt in c.output["covopts"]:
                 if self.covopts is None or covopt in self.covopts:
-                    self.files.append(c.output["base_dict"][covopt])
+                    self.cosmomc_input_files.append(c.output["base_dict"][covopt])
+                    self.cosmomc_output_files.append(c.output["label"] + covopt + ".csv.gz")
                     self.names.append(c.output["label"].replace("_", " ") + " " + covopt)
                     for p in c.output["cosmology_params"]:
                         if p not in self.params:
                             self.params.append(p)
-            self.logger.debug(f"Analyse task will create CosmoMC plots with {len(self.files)} covopts/plots")
+            self.logger.debug(f"Analyse task will create CosmoMC plots with {len(self.cosmomc_input_files)} covopts/plots")
 
         self.wsummary_files = [b.output["w_summary"] for b in self.biascor_deps]
 
+        # Get the fitres and m0diff files we'd want to parse for Hubble diagram plotting
+        self.biascor_fitres_input_files = [os.path.join(m, "SALT2mu_FITOPT000_MUOPT000.FITRES") for b in self.biascor_deps for m in b.output["m0dif_dirs"]]
+        self.biascor_prob_col_names = [b.output["prob_column_name"] for b in self.biascor_deps for m in b.output["m0dif_dirs"]]
+        self.biascor_fitres_output_files = [b.name + m.replace("SALT2mu_FITJOBS", 1) + "_FITOPT0_MUOPT0.fitres" for b in self.biascor_deps for m in b.output["m0dif_dirs"]]
+
+        # Get the m0diff files for everything
+        self.biascor_m0diffs = []
+        for b in self.biascor_deps:
+            for m in b.output["m0dif_dirs"]:
+                sim_number = 1
+                if os.path.basename(m).isdigit():
+                    sim_number = int(os.path.basename(m))
+                files = [f for f in sorted(os.listdir(m)) if f.endswith(".M0DIFF")]
+                for f in files:
+                    muopt_num = int(f.split("MUOPT")[-1].split(".")[0])
+                    fitopt_num = int(f.split("FITOPT")[-1].split("_")[0])
+                    if muopt_num == 0:
+                        muopt = "DEFAULT"
[…]
