+++ pippin/external/plot_biascor.py
@@ -31,17 +31,8 @@ def get_arguments():
     return config
 
 
-def load_file(file, args):
-    np.random.seed(123)
-    logging.info(f"Loading in file {file}")
-    if not os.path.exists(file):
-        logging.warning(f"File {file} not count, continuing and hoping this is executing on your local compute and you have the summary")
-        return None
+def load_file(file):
     df = pd.read_csv(file)
-    if "w" in args.get("BLIND", []):
-        df["w"] += np.random.normal(loc=0, scale=0.2, size=1000)[343]
-    if "om" in args.get("BLIND", []):
-        df["OM"] += np.random.normal(loc=0, scale=0.1, size=1000)[432]
     return df
 
 
@@ -59,26 +50,6 @@ def plot_single_file(source_file, df):
         c.add_covariance(means, cov, parameters=labels, name=f"Realisation {index}")
     c.plotter.plot_summary(errorbar=True, filename=output_file)
     del c
-
-
-def make_summary_file(wfit_files):
-    logging.info("Creating summary file")
-    all_output_csv = "all_biascor_individual.csv"
-
-    if os.path.exists(all_output_csv):
-        df_all = pd.read_csv(all_output_csv)
-    else:
-        df_all = None
-        for f in wfit_files:
-            df = pd.read_csv(f)
-            name = os.path.basename(os.path.dirname(os.path.dirname(f)))
-            df["name"] = name
-            if df_all is None:
-                df_all = df
-            else:
-                df_all = pd.concat([df_all, df])
-        df_all.to_csv(all_output_csv, index=False, float_format="%0.5f")
-    return df_all
 
 
 def plot_all_files(df_all):
@@ -174,22 +145,237 @@ def plot_scatter_comp(df_all):
         fig.savefig(figname, bbox_inches="tight", dpi=150, transparent=True)
 
 
+def make_hubble_plot(fitres_file, m0diff_file, prob_col_name, args):
+    logging.info("Making Hubble plot")
+    # Note that the fitres file has mu and fit 0, m0diff will have to select down to it
+
+    name, sim_num, *_ = fitres_file.split("_")
+    sim_num = int(sim_num)
+
+    df = pd.read_csv(fitres_file, delim_whitespace=True, comment="#")
+    dfm = pd.read_csv(m0diff_file)
+    dfm = dfm[(dfm.name == name) & (dfm.sim_num == sim_num) & (dfm.muopt_num == 0) & (dfm.fitopt_num == 0)]
+
+    from astropy.cosmology import FlatwCDM
+    import numpy as np
+    import matplotlib.pyplot as plt
+
+    df.sort_values(by="zHD", inplace=True)
+    dfm.sort_values(by="z", inplace=True)
+    dfm = dfm[dfm["MUDIFERR"] < 10]
+
+    ol = dfm.ol_ref.unique()[0]
+    w = dfm.w_ref.unique()[0]
+    alpha = 0
+    beta = 0
+    sigint = 0
+    gamma = r"$\gamma = 0$"
+    scalepcc = "NA"
+    num_sn_fit = df.shape[0]
+    contam_data, contam_true = "", ""
+
+    with open(fitres_file) as f:
+        for line in f.read().splitlines():
+            if "NSNFIT" in line:
+                v = int(line.split("=", 1)[1].strip())
+                num_sn_fit = v
+                num_sn = f"$N_{{SN}} = {v}$"
+            if "alpha0" in line and "=" in line and "+-" in line:
[…]
