+++ pippin/snana_fit.py
@@ -93,10 +93,13 @@ class SNANALightCurveFit(ConfigBasedExecutable):
                 self.fitopts.append(f)
         # Map the fitopt outputs
         mapped = {"DEFAULT": "FITOPT000.FITRES"}
+        mapped2 = {0: "DEFAULT"}
         for i, line in enumerate(self.fitopts):
             label = line.split("[")[1].split("]")[0]
             mapped[line] = f"FITOPT{i + 1:3d}.FITRES"
+            mapped2[i] = label
         self.output["fitopt_map"] = mapped
+        self.output["fitopt_index"] = mapped
         self.output["fitres_file"] = os.path.join(self.fitres_dirs[0], mapped["DEFAULT"])
 
         self.options = self.config.get("OPTS", {})

[…]
