### 2020-02-26 #2096 @danielskatz
This is a COI according to [JOSS policies](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy) but it's long enough ago and minor enough that it can be waived
