### 2020-02-26 #2096 @arfon
> In the interest of transparency, @Samreay and I worked at the same institution for a few months around three years ago. We did not work directly together on any projects, and I don't think that this is a COI that would make me unable to make an impartial evaluation.

:zap: thanks @kboone!
