### 2020-02-26 #2122 @whedon
[REVIEW]: Pippin: A pipeline for supernova cosmology
**Submitting author:** @samreay (<a href="http://orcid.org/0000-0003-2071-9349">Samuel Hinton</a>)
**Repository:** <a href="https://github.com/Samreay/Pippin" target ="_blank">https://github.com/Samreay/Pippin</a>
**Version:** 0.1.1

- [x] **Contribution and authorship:** Has the submitting author (samreay) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?

- [x] **Contribution and authorship:** Has the submitting author (samreay) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?
