### 2020-02-26 #2122 @arfon
@kboone, @temuller - please carry out your review in this issue by updating the checklist above and giving feedback in this issue. The reviewer guidelines are available here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html and if possible I'd appreciate it if you could complete your review in the next 2-3 weeks.

Any questions/concerns please let me know!
