### 2020-02-26 650d76a
Updating config to new syntax
