+++ configs/srh/srh_deslowz_errorbudget.yml
@@ -174,5 +174,5 @@ COSMOMC:
 ANALYSE:
   ALL_OMW:
     MASK_COSMOMC: OMW
-    HISTOGRAM: [DS_SIMDATADES5YR, D_DESSIMBIAS5YRIA_G10, D_DESSIMBIAS5YRCC, D_DESSIMBIAS5YRIA_C11]
+    MASK_LCFIT: [DS_SIMDATADES5YR, D_DESSIMBIAS5YRIA_G10, D_DESSIMBIAS5YRCC, D_DESSIMBIAS5YRIA_C11]
 
