### 2020-02-27 9c0881b
Updating config to new syntax
