### 2020-02-27 9b18338
Updating simulation to do typing at the start
