### 2020-02-27 2cbf8fc
Updating code formatting
