+++ pippin/analyse.py
@@ -111,7 +111,9 @@ class AnalyseChains(Task):  # TODO: Define the location of the output so we can 
         # Get the fitres and m0diff files we'd want to parse for Hubble diagram plotting
         self.biascor_fitres_input_files = [os.path.join(m, "SALT2mu_FITOPT000_MUOPT000.FITRES") for b in self.biascor_deps for m in b.output["m0dif_dirs"]]
         self.biascor_prob_col_names = [b.output["prob_column_name"] for b in self.biascor_deps for m in b.output["m0dif_dirs"]]
-        self.biascor_fitres_output_files = [b.name + "_" + m.replace("SALT2mu_FITJOBS", "1") + "_FITOPT0_MUOPT0.fitres" for b in self.biascor_deps for m in b.output["m0dif_dirs"]]
+        self.biascor_fitres_output_files = [
+            b.name + "_" + m.replace("SALT2mu_FITJOBS", "1") + "_FITOPT0_MUOPT0.fitres" for b in self.biascor_deps for m in b.output["m0dif_dirs"]
+        ]
 
         self.biascor_m0diffs = []
         self.biascor_m0diff_output = "all_biascor_m0diffs.csv"
