### 2020-02-27 32779ca
Fixing issue in SNANA Simulation
