### 2020-02-28 f930ce6
Reducing number of inference samples for bayesian model
