### 2020-02-28 ae911e0
Removing hubble plot copy code
