### 2020-02-28 96cd515
Updating plotting script
