### 2020-02-28 f28b04d
Updating plotting script
