### 2020-02-28 11d6e7a
Updating plotting script
