### 2020-02-28 aa6f1ab
Updating plotting script
