### 2020-02-28 e73a9cf
Updating plotting script
