+++ pippin/external/plot_errbudget.py
@@ -5,80 +5,19 @@ import sys
 import argparse
 import os
 import logging
-import seaborn as sns
-import matplotlib.pyplot as plt
 
 
-def load_params(file):
-    assert os.path.exists(file), f"Paramnames file {file} does not exist"
-    names, labels = [], []
-    with open(file) as f:
-        for line in f.read().splitlines():
-            n, l = line.split(maxsplit=1)
-            names.append(n.replace("*", ""))
-            labels.append(l)
-    return names, labels
-
-
-def load_chains(files, all_cols, use_cols=None):
-    header = ["weights", "likelihood"] + all_cols
-    data = [pd.read_csv(f, delim_whitespace=True, header=None, names=header) for f in files]
-
-    # Remove burn in by cutting off first 30%
-    data = [d.iloc[int(d.shape[0] * 0.3) :, :] for d in data]
-
-    combined = pd.concat(data)
-    if use_cols is None:
-        use_cols = all_cols
-    weights = combined["weights"].values
-    likelihood = combined["likelihood"].values
-    chain = combined[use_cols].values
-    return weights, likelihood, chain
-
-
-def getParamAndErr(name, param, chainfilebase):
-    logging.info("Loading in data from original CosmoMC files")
-    param_file = os.path.join(chainfilebase) + ".paramnames"
-    names, labels = load_params(param_file)
-    chain_files = get_chain_files(chainfilebase)
-    weights, likelihood, chain = load_chains(chain_files, names, use_cols=param)
-    return chain.mean(), chain.std()
+def setup_logging():
+    fmt = "[%(levelname)8s |%(filename)21s:%(lineno)3d]   %(message)s"
+    handler = logging.StreamHandler(sys.stdout)
+    logging.basicConfig(level=logging.DEBUG, format=fmt, handlers=[handler, logging.FileHandler("plot_errbudget.log")])
+    logging.getLogger("matplotlib").setLevel(logging.ERROR)
 
 
 def fail(msg, condition=True):
     if condition:
         logging.error(msg)
         raise ValueError(msg)
-
-
-def get_chain_files(basename):
-    folder = os.path.dirname(basename)
-    logging.info(f"Looking for chains in folder {folder}")
-    base = os.path.basename(basename)
-    files = [os.path.join(folder, f) for f in sorted(os.listdir(folder)) if base in f and f.endswith(".txt")]
-    fail(f"No chain files found for {os.path.join(folder, basename)}", condition=len(files) == 0)
-    logging.info(f"{len(files)} chains found for basename {basename}")
-    return files
-
-
-def setup_logging():
-    fmt = "[%(levelname)8s |%(filename)21s:%(lineno)3d]   %(message)s"
-    handler = logging.StreamHandler(sys.stdout)
-    logging.basicConfig(level=logging.DEBUG, format=fmt, handlers=[handler, logging.FileHandler("plot_cosmomc.log")])
-    logging.getLogger("matplotlib").setLevel(logging.ERROR)
-
-
-def blind(chain, names, columns_to_blind, index=0):
-    np.random.seed(123)
-    for i, c in enumerate(columns_to_blind):
-        logging.info(f"Blinding column {c}")
-        try:
-            ii = names.index(c)
-            scale = np.random.normal(loc=1, scale=0.0, size=1000)[321 + i]
-            offset = np.random.normal(loc=0, scale=0.2, size=1000)[343 + i + (index + 10)]
-            chain[:, ii] = chain[:, ii] * scale + np.std(chain[:, ii]) * offset
-        except ValueError as e:
-            logging.warning(f"Cannot find blinding column {c} in list of names {names}")
 
 
 def get_arguments():
[…]
