### 2020-02-28 8c50908
Fixing bug in parsed output files
