### 2020-02-28 8c1f68e
Fixing bug in parsed output files
