### 2020-02-28 b4b533c
Fixing bug in parsed output files
