+++ pippin/external/plot_errbudget.py
@@ -88,7 +88,7 @@ if __name__ == "__main__":
             # At this point, we have all the loaded in to a single dataframe, and now we group by name, compute the metrics, and save to file
             dfg = df_all.groupby("name")
             for name, df in dfg:
-                output_filename = f"errbudget_{name}.txt"
+                output_filename = f"errbudget_{name}.txt".replace(" ", "_")
 
                 avg_cols = [c for c in df.columns if c.endswith(" avg")]
                 std_cols = [c for c in df.columns if c.endswith(" std")]
