### 2020-03-02 cc553fe
Adding fitprob cut
