+++ pippin/external/compute_map.py
@@ -45,16 +45,16 @@ def parse_data(filename):
     mask_sbfluxcal = all_data["SBFLUXCAL"] > 0
     mask_5yr = all_data["MJD"] < 57850
     mask_nolowz_y1 = (all_data["MJD"] > 56800) | (all_data["zHD"] > 0.13)
-    mask_fitprob = all_data["FITPROB"] > 0.01
+    # mask_fitprob = all_data["FITPROB"] > 0.01
 
     # mask_ratio = all_data["RATIO"] > 0
-    mask = mask_errtest_high & mask_errtest_low & mask_sbfluxcal & mask_5yr & mask_nolowz_y1 & mask_fitprob  # & mask_fluxcalsim & mask_ratio
+    mask = mask_errtest_high & mask_errtest_low & mask_sbfluxcal & mask_5yr & mask_nolowz_y1  # & mask_fitprob  # & mask_fluxcalsim & mask_ratio
     print_drop(mask_errtest_high, "errtest < 10")
     print_drop(mask_errtest_low, "errtest >= 0.1")
     print_drop(mask_sbfluxcal, "SBFLUXCAL > 0")
     print_drop(mask_5yr, "MJD < 57850")
     print_drop(mask_nolowz_y1, "No Y1 z<0.13")
-    print_drop(mask_fitprob, "fitprob >0.01")
+    # print_drop(mask_fitprob, "fitprob >0.01")
     print_drop(mask, "Combined mask")
 
     all_data = all_data.loc[mask, :]
@@ -63,14 +63,14 @@ def parse_data(filename):
     print(f"Saving merged and thresholded data to {filename}")
     all_data.to_pickle(filename)
 
-    # Finally, lets get rid of massive outlier points. 15sigma should be good enough.
 
-
-def get_data(filename, cut):
+def get_data(filename, cut, cut_fitprob=0.0):
     df = pd.read_pickle(filename)
     mask2 = np.abs(df["DEVIATIONSIM"]) < cut
     print_drop(mask2, f"{cut} sigma outlier")
-    return df.loc[mask2, :]
+    mask_fitprob = df["FITPROB"] >= cut_fitprob
+    print_drop(mask_fitprob, f"FITPROB >= {cut_fitprob}")
+    return df.loc[mask2 & mask_fitprob, :]
 
 
 def rejstd(x, debug=False):
@@ -119,116 +119,117 @@ if not os.path.exists("fakes_obs.pkl"):
     print("Finished parsing")
 
 for cut in cuts:
-    print("Loading data")
-    df_f = get_data("fakes_obs.pkl", cut)
-    df_s = get_data("sim_obs.pkl", cut)
+    for cut_fitprob in [0.00, 0.01]:
+        print("Loading data")
+        df_f = get_data("fakes_obs.pkl", cut, cut_fitprob=cut_fitprob)
+        df_s = get_data("sim_obs.pkl", cut, cut_fitprob=cut_fitprob)
 
-    data = [
-        [("LOGSNR", np.arange(0, 3.1, 1)), ("SBMAG", np.arange(20, 30, 2)), ("PSF", np.array([1, 3, 5]))],
-        [("LOGSNR", np.arange(0, 3.1, 1)), ("SBMAG", np.arange(20, 30, 1))],
-        [("SBMAG", np.arange(20, 30, 1))],
-    ]
-    # data = [[("LOGSNR", np.arange(0, 3, 0.5)), ]]
-    # data = [[("SBMAG", np.arange(20, 30, 1))]]
+        data = [
+            [("LOGSNR", np.arange(0, 3.1, 1)), ("SBMAG", np.arange(20, 30, 2)), ("PSF", np.array([1, 3, 5]))],
[…]
