### 2020-03-03 a3dfcaa
Updating where PATH_USER_INPUT is set, and removing the FITPROB cut
