### 2020-03-03 86b9acd
Adding debug output for cteate_covariance
