### 2020-03-03 a3a6f14
Rerunning check maps with comparing the fitprob cut
