### 2020-03-03 15ab95f
Rerunning check maps with comparing the fitprob cut
