### 2020-03-03 423d4ce
Rerunning check maps with comparing the fitprob cut
