+++ configs/srh/srh_fakes_checkmaps.yml
@@ -235,34 +235,34 @@ LCFIT:
 ANALYSE:
 
   CHECK_FAKES:
-    HISTOGRAM:
+    MASK_LCFIT:
       - F1D_FAKES
       - F1D1_FAKES
       - F1D_SIMFAKE_1D_MIX_0
       - F1D_SIMFAKE_1D_MIX_1
 
   CHECK_FAKES_BEFORE_AFTER:
-    HISTOGRAM:
+    MASK_LCFIT:
       - F0D_FAKES
       - F1D_SIMFAKES_NOCORR
       - F1D_FAKES
       - F1D_SIMFAKE_1D_MIX
 
   CHECK_5YR_DIST_1D:
-    HISTOGRAM:
+    MASK_LCFIT:
       - D1D0_DATADES5YR
       - D1D1_DATADES5YR
       - D2D_SIMDES5YR_1D_REDCOV
       - D2D_SIMDES5YR_1D_INDCOV
 
   CHECK_5YR_DIST_2D:
-    HISTOGRAM:
+    MASK_LCFIT:
       - D2D_DATADES5YR
       - D2D_SIMDES5YR_2D_REDCOV
       - D2D_SIMDES5YR_2D_INDCOV
 
   CHECK_5YR_DIST_3D:
-    HISTOGRAM:
+    MASK_LCFIT:
       - D3D_DATADES5YR
       - D2D_SIMDES5YR_3D_REDCOV
       - D2D_SIMDES5YR_3D_INDCOV
