### 2020-03-03 68cfc42
Rerunning check maps with comparing the fitprob cut
