+++ configs/srh/srh_fakes_checkmaps.yml
@@ -93,7 +93,7 @@ SIM:
   SIMFAKE_1D_CORR_ALL00:
     IA: *fakesinput
     GLOBAL:
-      FLUXERRMODEL_FILE: $DES_ROOT/simlibs/maps/DES5YR_SIM_ERRORFUDGES_DIFFIMG_SBMAG_5.DAT
+      FLUXERRMODEL_FILE: $DES_ROOT/simlibs/maps/DES5YR_SIM_ERRORFUDGES_DIFFIMG_SBMAG_5_0.DAT
       BATCH_INFO: sbatch $SBATCH_TEMPLATES/SBATCH_Midway2_10hr.TEMPLATE 1
 
   SIMFAKE_2D_CORR_ALL00:
@@ -105,7 +105,7 @@ SIM:
   SIMFAKE_3D_CORR_ALL00:
     IA: *fakesinput
     GLOBAL:
-      FLUXERRMODEL_FILE: $DES_ROOT/simlibs/maps/DES5YR_SIM_ERRORFUDGES_DIFFIMG_LOGSNR_SBMAG_PSF_5.DAT
+      FLUXERRMODEL_FILE: $DES_ROOT/simlibs/maps/DES5YR_SIM_ERRORFUDGES_DIFFIMG_LOGSNR_SBMAG_PSF_5_0.DAT
       BATCH_INFO: sbatch $SBATCH_TEMPLATES/SBATCH_Midway2_10hr.TEMPLATE 1
 
 
@@ -134,7 +134,7 @@ SIM:
     <<: *deslikesim
     GLOBAL:
       <<: *default_corr
-      FLUXERRMODEL_FILE: $DES_ROOT/simlibs/maps/DES5YR_SIM_ERRORFUDGES_DIFFIMG_SBMAG_5.DAT
+      FLUXERRMODEL_FILE: $DES_ROOT/simlibs/maps/DES5YR_SIM_ERRORFUDGES_DIFFIMG_SBMAG_5_0.DAT
       RANSEED_REPEAT: 5 12345
       NGEN_UNIT: 1
       BATCH_INFO: sbatch $SBATCH_TEMPLATES/SBATCH_Midway2_10hr.TEMPLATE 5
@@ -142,7 +142,7 @@ SIM:
   SIMDES5YR_1D_INDCOV:
     <<: *deslikesim
     GLOBAL:
-      FLUXERRMODEL_FILE: $DES_ROOT/simlibs/maps/DES5YR_SIM_ERRORFUDGES_DIFFIMG_SBMAG_5.DAT
+      FLUXERRMODEL_FILE: $DES_ROOT/simlibs/maps/DES5YR_SIM_ERRORFUDGES_DIFFIMG_SBMAG_5_0.DAT
       RANSEED_REPEAT: 5 12345
       NGEN_UNIT: 1
       BATCH_INFO: sbatch $SBATCH_TEMPLATES/SBATCH_Midway2_10hr.TEMPLATE 5
@@ -151,7 +151,7 @@ SIM:
     <<: *deslikesim
     GLOBAL:
       <<: *default_corr
-      FLUXERRMODEL_FILE: $DES_ROOT/simlibs/maps/DES5YR_SIM_ERRORFUDGES_DIFFIMG_LOGSNR_SBMAG_PSF_5.DAT
+      FLUXERRMODEL_FILE: $DES_ROOT/simlibs/maps/DES5YR_SIM_ERRORFUDGES_DIFFIMG_LOGSNR_SBMAG_PSF_5_0.DAT
       RANSEED_REPEAT: 5 12345
       NGEN_UNIT: 1
[…]
