+++ configs/srh/srh_fakes_sanitycheck.yml
@@ -24,24 +24,42 @@ SIM:
       FLUXERRMODEL_FILE: NONE
       RANSEED_REPEAT: 1 12345
 
-  SIMFAKE_1D_CORRECTION:
+  SIMFAKE_1D_CORRECTION_F0:
     IA: *fakesinput
     GLOBAL:
-      FLUXERRMODEL_FILE: $DES_ROOT/simlibs/maps/DES5YR_SIM_ERRORFUDGES_DIFFIMG_SBMAG_5.DAT
+      FLUXERRMODEL_FILE: $DES_ROOT/simlibs/maps/DES5YR_SIM_ERRORFUDGES_DIFFIMG_SBMAG_5_0.DAT
       RANSEED_REPEAT: 1 12345
 
-  SIMFAKE_2D_CORRECTION:
+  SIMFAKE_1D_CORRECTION_F1:
     IA: *fakesinput
     GLOBAL:
-      FLUXERRMODEL_FILE: $DES_ROOT/simlibs/maps/DES5YR_SIM_ERRORFUDGES_DIFFIMG_LOGSNR_SBMAG_5.DAT
+      FLUXERRMODEL_FILE: $DES_ROOT/simlibs/maps/DES5YR_SIM_ERRORFUDGES_DIFFIMG_SBMAG_5_1.DAT
       RANSEED_REPEAT: 1 12345
 
-  SIMFAKE_3D_CORRECTION:
+  SIMFAKE_2D_CORRECTION_F0:
+    IA: *fakesinput
+    GLOBAL:
+      FLUXERRMODEL_FILE: $DES_ROOT/simlibs/maps/DES5YR_SIM_ERRORFUDGES_DIFFIMG_LOGSNR_SBMAG_5_0.DAT
+      RANSEED_REPEAT: 1 12345
+
+  SIMFAKE_2D_CORRECTION_F1:
+    IA: *fakesinput
+    GLOBAL:
+      FLUXERRMODEL_FILE: $DES_ROOT/simlibs/maps/DES5YR_SIM_ERRORFUDGES_DIFFIMG_LOGSNR_SBMAG_5_1.DAT
+      RANSEED_REPEAT: 1 12345
+
+
+  SIMFAKE_3D_CORRECTION_F0:
     IA: *fakesinput
     GLOBAL:
       FLUXERRMODEL_FILE: $DES_ROOT/simlibs/maps/DES5YR_SIM_ERRORFUDGES_DIFFIMG_LOGSNR_SBMAG_PSF_5.DAT
       RANSEED_REPEAT: 1 12345
 
+  SIMFAKE_3D_CORRECTION_F1:
+    IA: *fakesinput
+    GLOBAL:
+      FLUXERRMODEL_FILE: $DES_ROOT/simlibs/maps/DES5YR_SIM_ERRORFUDGES_DIFFIMG_LOGSNR_SBMAG_PSF_5.DAT
+      RANSEED_REPEAT: 1 12345
 
 LCFIT:
 
@@ -52,46 +70,74 @@ LCFIT:
       FLUXERRMODEL_FILE: NONE
     OPTS: *default_fitopts
 
-  F1D:
+  F1D0:
     BASE: surveys/des/lcfit_nml/fakes.nml
-    MASK: [DATAFAKES, SIMFAKE_1D_CORRECTION]
+    MASK: [DATAFAKES, SIMFAKE_1D_CORRECTION_F0]
     SNLCINP:
-      FLUXERRMODEL_FILE: $DES_ROOT/simlibs/maps/DES5YR_FAKES_ERRORFUDGES_DIFFIMG_SBMAG_5.DAT
[…]
