+++ configs/srh/srh_fakes_sanitycheck.yml
@@ -115,28 +115,28 @@ LCFIT:
 ANALYSE:
 
   SANITY_0D:
-    HISTOGRAM:
+    MASK_LCFIT:
       - F0D0_DATAFAKES
       - F0D0_SIMFAKE_NOCORRECTION_F0
       - F0D1_DATAFAKES
       - F0D1_SIMFAKE_NOCORRECTION_F1
 
   SANITY_1D:
-    HISTOGRAM:
+    MASK_LCFIT:
       - F1D0_DATAFAKES
       - F1D0_SIMFAKE_1D_CORRECTION_F0
       - F1D1_DATAFAKES
       - F1D1_SIMFAKE_1D_CORRECTION_F1
 
   SANITY_2D:
-    HISTOGRAM:
+    MASK_LCFIT:
       - F2D0_DATAFAKES
       - F2D0_SIMFAKE_2D_CORRECTION_F0
       - F2D1_DATAFAKES
       - F2D1_SIMFAKE_2D_CORRECTION_F1
 
   SANITY_3D:
-    HISTOGRAM:
+    MASK_LCFIT:
       - F3D0_DATAFAKES
       - F3D0_SIMFAKE_3D_CORRECTION_F0
       - F3D1_DATAFAKES
