+++ configs/srh/srh_fakes_sanitycheck.yml
@@ -116,10 +116,8 @@ ANALYSE:
 
   SANITY_0D:
     MASK_LCFIT:
-      - F0D0_DATAFAKES
-      - F0D0_SIMFAKE_NOCORRECTION_F0
-      - F0D1_DATAFAKES
-      - F0D1_SIMFAKE_NOCORRECTION_F1
+      - F0D_DATAFAKES
+      - F0D_SIMFAKE_NOCORRECTION
 
   SANITY_1D:
     MASK_LCFIT:
