+++ configs/srh/srh_fakes_sanitycheck.yml
@@ -52,13 +52,13 @@ SIM:
   SIMFAKE_3D_CORRECTION_F0:
     IA: *fakesinput
     GLOBAL:
-      FLUXERRMODEL_FILE: $DES_ROOT/simlibs/maps/DES5YR_SIM_ERRORFUDGES_DIFFIMG_LOGSNR_SBMAG_PSF_5.DAT
+      FLUXERRMODEL_FILE: $DES_ROOT/simlibs/maps/DES5YR_SIM_ERRORFUDGES_DIFFIMG_LOGSNR_SBMAG_PSF_5_0.DAT
       RANSEED_REPEAT: 1 12345
 
   SIMFAKE_3D_CORRECTION_F1:
     IA: *fakesinput
     GLOBAL:
-      FLUXERRMODEL_FILE: $DES_ROOT/simlibs/maps/DES5YR_SIM_ERRORFUDGES_DIFFIMG_LOGSNR_SBMAG_PSF_5.DAT
+      FLUXERRMODEL_FILE: $DES_ROOT/simlibs/maps/DES5YR_SIM_ERRORFUDGES_DIFFIMG_LOGSNR_SBMAG_PSF_5_1.DAT
       RANSEED_REPEAT: 1 12345
 
 LCFIT:
