### 2020-03-03 af26b37
Shrinking axis labels as they dont want to hide
