+++ configs/srh/srh_fakes_getcov.yml
@@ -11,12 +11,16 @@
 # Oh btw, I used YAML anchoring and aliasing in here. How good is it. Just adding it here so its easy to see
 ANCHORS:  # The name of this doesnt actually matter, its all about the & and *
 
-  DEFAULT_SIM_MAP3D: &map3d
-    FLUXERRMODEL_FILE: $DES_ROOT/simlibs/maps/DES5YR_SIM_ERRORFUDGES_DIFFIMG_LOGSNR_SBMAG_PSF_5.DAT
+  DEFAULT_SIM_MAP1D: &map1d
+    FLUXERRMODEL_FILE: $DES_ROOT/simlibs/maps/DES5YR_SIM_ERRORFUDGES_DIFFIMG_SBMAG_5_0.DAT
     RANSEED_REPEAT: 1 12345
 
   DEFAULT_SIM_MAP2D: &map2d
-    FLUXERRMODEL_FILE: $DES_ROOT/simlibs/maps/DES5YR_SIM_ERRORFUDGES_DIFFIMG_LOGSNR_SBMAG_5.DAT
+    FLUXERRMODEL_FILE: $DES_ROOT/simlibs/maps/DES5YR_SIM_ERRORFUDGES_DIFFIMG_LOGSNR_SBMAG_5_0.DAT
+    RANSEED_REPEAT: 1 12345
+
+  DEFAULT_SIM_MAP3D: &map3d
+    FLUXERRMODEL_FILE: $DES_ROOT/simlibs/maps/DES5YR_SIM_ERRORFUDGES_DIFFIMG_LOGSNR_SBMAG_PSF_5_0.DAT
     RANSEED_REPEAT: 1 12345
 
   FAKES_INPUT: &fakesinput
@@ -26,10 +30,15 @@ ANCHORS:  # The name of this doesnt actually matter, its all about the & and *
     BATCH_INFO: sbatch $SBATCH_TEMPLATES/SBATCH_Midway2_10hr.TEMPLATE 5
     APPEND_TABLE_TEXT: SNRMAX_g SNRMAX_r SNRMAX_i SNRMAX_z chi2_g chi2_r chi2_i chi2_z
 
-  SNLCINP3D: &map3ddata
-    FLUXERRMODEL_FILE: $DES_ROOT/simlibs/maps/DES5YR_FAKES_ERRORFUDGES_DIFFIMG_LOGSNR_SBMAG_PSF_5.DAT
+  SNLCINP1D: &map1ddata
+    FLUXERRMODEL_FILE: $DES_ROOT/simlibs/maps/DES5YR_FAKES_ERRORFUDGES_DIFFIMG_SBMAG_5_0.DAT
+
   SNLCINP2D: &map2ddata
-    FLUXERRMODEL_FILE: $DES_ROOT/simlibs/maps/DES5YR_FAKES_ERRORFUDGES_DIFFIMG_LOGSNR_SBMAG_5.DAT
+    FLUXERRMODEL_FILE: $DES_ROOT/simlibs/maps/DES5YR_FAKES_ERRORFUDGES_DIFFIMG_LOGSNR_SBMAG_5_0.DAT
+
+  SNLCINP3D: &map3ddata
+    FLUXERRMODEL_FILE: $DES_ROOT/simlibs/maps/DES5YR_FAKES_ERRORFUDGES_DIFFIMG_LOGSNR_SBMAG_PSF_5_0.DAT
+
 
 DATAPREP:
   FAKES:
@@ -122,8 +131,6 @@ SIM:
       <<: *map2d
       FLUXERRMODEL_REDCOV: "griz:0.4"
 
-
-
   SIMFAKE_2D_CORR_INDEP01:
     IA: *fakesinput
     GLOBAL:
@@ -146,6 +153,51 @@ SIM:
[…]
