### 2020-03-04 1c7a4f2
Shrinking font size down for the histogram
