+++ pippin/external/plot_histogram.py
@@ -138,14 +138,7 @@ def plot_histograms(data, sims, types, figname):
                 red_chi2 = chi2 / ndof
 
                 ax.text(
-                    0.99,
-                    0.99 - 0.1 * i,
-                    f"{chi2:0.1f}/{ndof:d}={red_chi2:0.1f}",
-                    horizontalalignment="right",
-                    verticalalignment="top",
-                    transform=ax.transAxes,
-                    color=cs[i],
-                    fontsize=10,
+                    0.99, 0.99 - 0.1 * i, f"{chi2:0.1f}/{ndof:d}={red_chi2:0.1f}", horizontalalignment="right", verticalalignment="top", transform=ax.transAxes, color=cs[i], fontsize=8,
                 )
 
     handles, labels = ax.get_legend_handles_labels()
@@ -166,12 +159,7 @@ def get_means_and_errors(x, y, bins):
 
     std, *_ = binned_statistic(x, y, bins=bins, statistic=lambda x: np.std(x))
     std_err, *_ = binned_statistic(
-        x,
-        y,
-        bins=bins,
-        statistic=lambda x: np.nan
-        if x.size < 20
-        else np.sqrt((1 / x.size) * (moment(x, 4) - (((x.size - 3) / (x.size - 1)) * np.var(x) ** 2))) / (2 * np.std(x)),
+        x, y, bins=bins, statistic=lambda x: np.nan if x.size < 20 else np.sqrt((1 / x.size) * (moment(x, 4) - (((x.size - 3) / (x.size - 1)) * np.var(x) ** 2))) / (2 * np.std(x)),
     )
     return means, err, std, std_err
 
