+++ configs/srh/srh_fakes_getcov.yml
@@ -296,13 +296,13 @@ ANALYSE:
   CORR_1D_05:
     MASK_LCFIT:
       - F1D_FAKES
-      - F3D_SIMFAKE_1D_CORR_ALL04
-      - F3D_SIMFAKE_1D_CORR_INDEP04
+      - F3D_SIMFAKE_1D_CORR_ALL05
+      - F3D_SIMFAKE_1D_CORR_INDEP05
   CORR_1D_06:
     MASK_LCFIT:
       - F1D_FAKES
-      - F3D_SIMFAKE_1D_CORR_ALL04
-      - F3D_SIMFAKE_1D_CORR_INDEP04
+      - F3D_SIMFAKE_1D_CORR_ALL06
+      - F3D_SIMFAKE_1D_CORR_INDEP06
 
   ALL_1D_CORRELATIONS:
     MASK_LCFIT:
