+++ configs/srh/srh_fakes_checkmaps.yml
@@ -11,13 +11,21 @@
 # Oh btw, I used YAML anchoring and aliasing in here. How good is it. Just adding it here so its easy to see
 ANCHORS:  # The name of this doesnt actually matter, its all about the & and *
 
-  DEFAULT_SIM_MAP: &default_map
-    FLUXERRMODEL_FILE: $DES_ROOT/simlibs/maps/DES5YR_SIM_ERRORFUDGES_DIFFIMG_SBMAG_5_0.DAT
-    RANSEED_REPEAT: 1 12345
 
-  DEFAULT_CORR: &default_corr
-    FLUXERRMODEL_REDCOV(DEEP): "g:0.05,r:0.05,i:0.05,z:0.05"
-    FLUXERRMODEL_REDCOV(SHALLOW): "griz:0.25"
+  DEFAULT_CORR1: &default_corr_1d
+    FLUXERRMODEL_FILE: $DES_ROOT/simlibs/maps/DES5YR_SIM_ERRORFUDGES_DIFFIMG_SBMAG_5_0.DAT
+    FLUXERRMODEL_REDCOV(SHALLOW): "griz:0.45"
+    FLUXERRMODEL_REDCOV(DEEP): "g:0.2,r:0.2,i:0.2,z:0.2"
+
+  DEFAULT_CORR2: &default_corr_2d
+    FLUXERRMODEL_FILE: $DES_ROOT/simlibs/maps/DES5YR_SIM_ERRORFUDGES_DIFFIMG_LOGSNR_SBMAG_5_0.DAT
+    FLUXERRMODEL_REDCOV(SHALLOW): "griz:0.4"
+    FLUXERRMODEL_REDCOV(DEEP): "g:0.15,r:0.15,i:0.15,z:0.15"
+
+  DEFAULT_CORR3: &default_corr_3d
+    FLUXERRMODEL_FILE: $DES_ROOT/simlibs/maps/DES5YR_SIM_ERRORFUDGES_DIFFIMG_LOGSNR_SBMAG_PSF_5_0.DAT
+    FLUXERRMODEL_REDCOV(SHALLOW): "griz:0.33"
+    FLUXERRMODEL_REDCOV(DEEP): "g:0.2,r:0.2,i:0.2,z:0.2"
 
   FAKES_INPUT: &fakesinput
     BASE: surveys/des/sims_ia/fakes_diffimg.input
@@ -26,8 +34,6 @@ ANCHORS:  # The name of this doesnt actually matter, its all about the & and *
     BATCH_INFO: sbatch $SBATCH_TEMPLATES/SBATCH_Midway2_10hr.TEMPLATE 10
     APPEND_TABLE_TEXT: SNRMAX_g SNRMAX_r SNRMAX_i SNRMAX_z chi2_g chi2_r chi2_i chi2_z
 
-  SNLCINP: &default_mapdata
-    FLUXERRMODEL_FILE: $DES_ROOT/simlibs/maps/DES5YR_FAKES_ERRORFUDGES_DIFFIMG_SBMAG_5_0.DAT
 
   DESLIKESIM: &deslikesim
     IA_G10:
@@ -71,98 +77,53 @@ SIM:
 
   # Great, so that should allow a good mix for the correlation to be determined
   # broken down by deep and shallow fields. Lets simulate this mix
-  SIMFAKE_1D_MIX_0:
+  SIMFAKE_1D_MIX:
     IA: *fakesinput
     GLOBAL:
-      <<: *default_map
-      <<: *default_corr
+      <<: *default_corr_1d
       BATCH_INFO: sbatch $SBATCH_TEMPLATES/SBATCH_Midway2_10hr.TEMPLATE 1
-
-  SIMFAKE_1D_MIX_1:
-    IA:
[…]
