### 2020-03-04 6b37099
Only generatring 5 sigma clipping
