### 2020-03-04 a7e4b07
Only generatring 5 sigma clipping
