+++ configs/srh/srh_fakes_checkmaps.yml
@@ -110,6 +110,7 @@ SIM:
       <<: *default_corr_1d
       NGEN_UNIT: 1
       RANSEED_REPEAT: 5 12345
+      BATCH_INFO: sbatch $SBATCH_TEMPLATES/SBATCH_Midway2_10hr.TEMPLATE 1
 
   SIMDES5YR_2D:
     <<: *deslikesim
@@ -117,6 +118,7 @@ SIM:
       <<: *default_corr_2d
       NGEN_UNIT: 1
       RANSEED_REPEAT: 5 12345
+      BATCH_INFO: sbatch $SBATCH_TEMPLATES/SBATCH_Midway2_10hr.TEMPLATE 1
 
   SIMDES5YR_3D:
     <<: *deslikesim
@@ -124,6 +126,7 @@ SIM:
       <<: *default_corr_3d
       NGEN_UNIT: 1
       RANSEED_REPEAT: 5 12345
+      BATCH_INFO: sbatch $SBATCH_TEMPLATES/SBATCH_Midway2_10hr.TEMPLATE 1
 
 
 
