### 2020-03-04 3fe3b12
Only generatring 5 sigma clipping
