### 2020-03-04 ff7797b
Updating scanning for logs
