### 2020-03-04 c4c057c
Updating scanning for logs
