+++ configs/srh/srh_fakes_checkmaps.yml
@@ -147,7 +147,7 @@ LCFIT:
     MASK: FAKE
     SNLCINP:
       FLUXERRMODEL_FILE: $DES_ROOT/simlibs/maps/DES5YR_FAKES_ERRORFUDGES_DIFFIMG_SBMAG_5_0.DAT
-      OPTS: *default_fitopts
+    OPTS: *default_fitopts
 
   F2D: # fit fakes using the 2D correction, which only apples to the real fakes, not the sims
     BASE: surveys/des/lcfit_nml/fakes.nml
