### 2020-03-04 3622454
Updating scanning for logs
