### 2020-03-04 b0bb9d5
Updating scanning for logs
