### 2020-03-05 5458470
Updating scanning for logs
