+++ configs/srh/srh_fakes_checkmaps.yml
@@ -108,7 +108,7 @@ SIM:
     <<: *deslikesim
     GLOBAL:
       <<: *default_corr_1d
-      NGEN_UNIT: 0.01
+      NGEN_UNIT: 1
       RANSEED_REPEAT: 1 12345
       BATCH_INFO: sbatch $SBATCH_TEMPLATES/SBATCH_Midway2_10hr.TEMPLATE 1
 
@@ -116,7 +116,7 @@ SIM:
     <<: *deslikesim
     GLOBAL:
       <<: *default_corr_2d
-      NGEN_UNIT: 0.01
+      NGEN_UNIT: 1
       RANSEED_REPEAT: 1 12345
       BATCH_INFO: sbatch $SBATCH_TEMPLATES/SBATCH_Midway2_10hr.TEMPLATE 1
 
@@ -124,7 +124,7 @@ SIM:
     <<: *deslikesim
     GLOBAL:
       <<: *default_corr_3d
-      NGEN_UNIT: 0.01
+      NGEN_UNIT: 1
       RANSEED_REPEAT: 1 12345
       BATCH_INFO: sbatch $SBATCH_TEMPLATES/SBATCH_Midway2_10hr.TEMPLATE 1
 
