### 2020-03-05 df97282
Updating scanning for logs
