### 2020-03-05 0e25472
Updating scanning for logs
