+++ configs/srh/srh_fakes_checkmaps.yml
@@ -71,8 +71,8 @@ SIM:
       BASE: surveys/des/sims_ia/fakes_diffimg.input
     GLOBAL:
       FLUXERRMODEL_FILE: NONE
-      RANSEED_REPEAT: 1 12345
-      BATCH_INFO: sbatch $SBATCH_TEMPLATES/SBATCH_Midway2_10hr.TEMPLATE 1
+      RANSEED_REPEAT: 5 12345
+      BATCH_INFO: sbatch $SBATCH_TEMPLATES/SBATCH_Midway2_10hr.TEMPLATE 5
 
 
   # Great, so that should allow a good mix for the correlation to be determined
