+++ configs/srh/srh_fakes_genmaps.yml
@@ -7,8 +7,8 @@
 #    sim obs dump file -> sim_obs.fitres
 # 3. Use these files to run compute_map.py   in the external code directory
 # 4. Copy the maps it produces into the $DES_ROOT/simlibs/maps directory
-# 5. Using the 1D 2D and 3D maps, run a sanity check using srh_fakes_sanitycheck.yml
-# 6. Using one of the maps (ie the 2D 5sigma map) as a fiducial map, run srh_fakes_getcov.yml to determine the covariance
+# 5. Using the 1D, 2D and 3D maps, run a sanity check using srh_fakes_sanitycheck.yml
+# 6. srh_fakes_getcov.yml to determine the covariance. The amount will vary slightly between the 1D, 2D, and 3D maps
 # 7. Now we need to check the scatter model impact. Hopefully this doesnt change, run srh_fakes_getphasecor.yml to determine best scatter
 # 8. Put those values into the srh_fakes_checkmaps.yml file to verify that it looks good on fakes and data
 
