+++ configs/srh/srh_fakes_getphasecor.yml
@@ -200,7 +200,7 @@ LCFIT:
 ANALYSE:
 
   CHECK_5YR_02:
-    HISTOGRAM:
+    MASK_LCFIT:
       - D1D_DATADES5YR
       - D1D_SIMDES5YR_NOPHASE
       - D1D_SIMDES5YR_02_0
@@ -210,7 +210,7 @@ ANALYSE:
       - D1D_SIMDES5YR_02_10
 
   CHECK_5YR_03:
-    HISTOGRAM:
+    MASK_LCFIT:
       - D1D_DATADES5YR
       - D1D_SIMDES5YR_NOPHASE
       - D1D_SIMDES5YR_03_0
@@ -220,7 +220,7 @@ ANALYSE:
       - D1D_SIMDES5YR_03_10
 
   CHECK_5YR_04:
-    HISTOGRAM:
+    MASK_LCFIT:
       - D1D_DATADES5YR
       - D1D_SIMDES5YR_NOPHASE
       - D1D_SIMDES5YR_04_0
@@ -230,7 +230,7 @@ ANALYSE:
       - D1D_SIMDES5YR_04_10
 
   CHECK_5YR_05:
-    HISTOGRAM:
+    MASK_LCFIT:
       - D1D_DATADES5YR
       - D1D_SIMDES5YR_NOPHASE
       - D1D_SIMDES5YR_05_0
