### 2020-03-05 6b4edb0
Updating snana_fit to use the newer methods
