### 2020-03-05 bc5a1dd
Updating snana_fit to use the newer methods
