### 2020-03-05 83c9bf3
Updating snana_fit to use the newer methods
