+++ configs/srh/srh_fakes_getphasecor.yml
@@ -16,7 +16,7 @@ ANCHORS:  # The name of this doesnt actually matter, its all about the & and *
       RANSEED_REPEAT: 5 12345
       FLUXERRMODEL_REDCOV(DEEP): "g:0.2,r:0.2,i:0.2,z:0.2"
       FLUXERRMODEL_REDCOV(SHALLOW): "griz:0.45"
-      SEARCHEFF_zHOST_FILE: $DES_ROOT/models/searcheff/samtest/efficiency_zHD_G10.dat
+      SEARCHEFF_zHOST_FILE: $DES_ROOT/models/searcheff/5yr/efficiency_zHD_G10.dat
       NGEN_UNIT: 1
 
 
