### 2020-03-05 68d1b34
Updating default amount of phase correlation
