### 2020-03-05 dae08e8
Updating default amount of phase correlation
