### 2020-03-06 803a67c
Updating default amount of phase correlation
