+++ configs/srh/srh_fakes_getphasecor.yml
@@ -12,7 +12,7 @@ ANCHORS:  # The name of this doesnt actually matter, its all about the & and *
     CC_TEMPLATES:
       BASE: surveys/des/sims_cc/sn_collection_maria_dust1.input
     GLOBAL:
-      FLUXERRMODEL_FILE: $DES_ROOT/simlibs/maps/DES5YR_SIM_ERRORFUDGES_DIFFIMG_SBMAG_5.DAT
+      FLUXERRMODEL_FILE: $DES_ROOT/simlibs/maps/DES5YR_SIM_ERRORFUDGES_DIFFIMG_SBMAG.DAT
       RANSEED_REPEAT: 5 12345
       FLUXERRMODEL_REDCOV(DEEP): "g:0.2,r:0.2,i:0.2,z:0.2"
       FLUXERRMODEL_REDCOV(SHALLOW): "griz:0.45"
@@ -178,7 +178,7 @@ LCFIT:
     BASE: surveys/des/lcfit_nml/des_5yr.nml
     MASK: DES
     SNLCINP:
-      FLUXERRMODEL_FILE: $DES_ROOT/simlibs/maps/DES5YR_FAKES_ERRORFUDGES_DIFFIMG_SBMAG_5.DAT
+      FLUXERRMODEL_FILE: $DES_ROOT/simlibs/maps/DES5YR_FAKES_ERRORFUDGES_DIFFIMG_SBMAG.DAT
       ROOTFILE_OUT: FIT.ROOT
     OPTS:
       BATCH_INFO: sbatch $SBATCH_TEMPLATES/SBATCH_Midway2_10hr.TEMPLATE 5
@@ -188,7 +188,7 @@ LCFIT:
     BASE: surveys/des/lcfit_nml/des_5yr.nml
     MASK: [DATADES5YR, SIMDES5YR_05_1, SIMDES5YR_NOPHASE]
     SNLCINP:
-      FLUXERRMODEL_FILE: $DES_ROOT/simlibs/maps/DES5YR_FAKES_ERRORFUDGES_DIFFIMG_SBMAG_5.DAT
+      FLUXERRMODEL_FILE: $DES_ROOT/simlibs/maps/DES5YR_FAKES_ERRORFUDGES_DIFFIMG_SBMAG.DAT
       ROOTFILE_OUT: FIT.ROOT
       SNTABLE_LIST: FITRES FITRES+RESIDUALS LCPLOT(text:csv)
     OPTS:
