### 2020-03-06 27699b2
Updating default amount of phase correlation
