+++ configs/srh/srh_fakes_getphasecor.yml
@@ -14,8 +14,6 @@ ANCHORS:  # The name of this doesnt actually matter, its all about the & and *
     GLOBAL:
       FLUXERRMODEL_FILE: $DES_ROOT/simlibs/maps/DES5YR_SIM_ERRORFUDGES_DIFFIMG_SBMAG.DAT
       RANSEED_REPEAT: 5 12345
-      FLUXERRMODEL_REDCOV(DEEP): "g:0.2,r:0.2,i:0.2,z:0.2"
-      FLUXERRMODEL_REDCOV(SHALLOW): "griz:0.45"
       SEARCHEFF_zHOST_FILE: $DES_ROOT/models/searcheff/5yr/efficiency_zHD_G10.dat
       NGEN_UNIT: 1
 
