### 2020-03-06 36f96d2
Updating default amount of phase correlation
