### 2020-03-06 047e1e1
Updating default amount of phase correlation
