+++ configs/srh/srh_fakes_checkmaps.yml
@@ -13,19 +13,13 @@ ANCHORS:  # The name of this doesnt actually matter, its all about the & and *
 
 
   DEFAULT_CORR1: &default_corr_1d
-    FLUXERRMODEL_FILE: $DES_ROOT/simlibs/maps/DES5YR_SIM_ERRORFUDGES_DIFFIMG_SBMAG_5.DAT
-    FLUXERRMODEL_REDCOV(SHALLOW): "griz:0.45"
-    FLUXERRMODEL_REDCOV(DEEP): "g:0.2,r:0.2,i:0.2,z:0.2"
+    FLUXERRMODEL_FILE: $DES_ROOT/simlibs/maps/DES5YR_SIM_ERRORFUDGES_DIFFIMG_SBMAG.DAT
 
   DEFAULT_CORR2: &default_corr_2d
-    FLUXERRMODEL_FILE: $DES_ROOT/simlibs/maps/DES5YR_SIM_ERRORFUDGES_DIFFIMG_LOGSNR_SBMAG_5.DAT
-    FLUXERRMODEL_REDCOV(SHALLOW): "griz:0.4"
-    FLUXERRMODEL_REDCOV(DEEP): "g:0.15,r:0.15,i:0.15,z:0.15"
+    FLUXERRMODEL_FILE: $DES_ROOT/simlibs/maps/DES5YR_SIM_ERRORFUDGES_DIFFIMG_LOGSNR_SBMAG.DAT
 
   DEFAULT_CORR3: &default_corr_3d
-    FLUXERRMODEL_FILE: $DES_ROOT/simlibs/maps/DES5YR_SIM_ERRORFUDGES_DIFFIMG_LOGSNR_SBMAG_PSF_5.DAT
-    FLUXERRMODEL_REDCOV(SHALLOW): "griz:0.33"
-    FLUXERRMODEL_REDCOV(DEEP): "g:0.2,r:0.2,i:0.2,z:0.2"
+    FLUXERRMODEL_FILE: $DES_ROOT/simlibs/maps/DES5YR_SIM_ERRORFUDGES_DIFFIMG_LOGSNR_SBMAG_PSF.DAT
 
   FAKES_INPUT: &fakesinput
     BASE: surveys/des/sims_ia/fakes_diffimg.input
@@ -146,21 +140,21 @@ LCFIT:
     BASE: surveys/des/lcfit_nml/fakes.nml
     MASK: FAKE
     SNLCINP:
-      FLUXERRMODEL_FILE: $DES_ROOT/simlibs/maps/DES5YR_FAKES_ERRORFUDGES_DIFFIMG_SBMAG_5.DAT
+      FLUXERRMODEL_FILE: $DES_ROOT/simlibs/maps/DES5YR_FAKES_ERRORFUDGES_DIFFIMG_SBMAG.DAT
     OPTS: *default_fitopts
 
   F2D: # fit fakes using the 2D correction, which only apples to the real fakes, not the sims
     BASE: surveys/des/lcfit_nml/fakes.nml
     MASK: FAKES
     SNLCINP:
-      FLUXERRMODEL_FILE: $DES_ROOT/simlibs/maps/DES5YR_FAKES_ERRORFUDGES_DIFFIMG_LOGSNR_SBMAG_5.DAT
+      FLUXERRMODEL_FILE: $DES_ROOT/simlibs/maps/DES5YR_FAKES_ERRORFUDGES_DIFFIMG_LOGSNR_SBMAG.DAT
     OPTS: *default_fitopts
 
   F3D: # fit fakes using the 2D correction, which only apples to the real fakes, not the sims
[…]
