### 2020-03-10 #2122 @temuller
See my review attached.
[JOSS_Review_2122.pdf](https://github.com/openjournals/joss-reviews/files/4312102/JOSS_Review_2122.pdf)
