### 2020-03-10 ffb6170
Adding link to SNANA and stating purpose of software clearer in main readme.
