### 2020-03-10 #2122 @Samreay
Hey Tomás! I've added a link to the SNANA report and made the purpose statement clearer in the main readme. As to reporting issues or seeking support, I've updated the contribution section to make it clear that it applies to seeking support and raising issues.

As to the functionality, this is definitely a hassle - because Pippin is a pipeline that stitches other software together, running it necessarily requires having those pieces installed (which is a big hassle for SNANA if you dont have a midway account). Even the data prep stages uses new code in SNANA to determine better starting conditions for the eventual light curve fitting of data.

I'd be happy to record a video showing the usage in action if that would help, just let me know.
