### 2020-03-10 bae4a6a
Making it clear the examples directory are examples of task configurations, not pipelines you can run
