+++ examples/sim.yml
@@ -1,3 +1,7 @@
+# Note that these examples are for the specific task only, and you cannot run this file (as there are no dependent tasks).
+# Use these files to see how you can configure your tasks, and the configs directory contains full analyses showing
+# all the tasks and how they relate to each other
+
 SIM:
   LOWZSIM:
     IA_G10_LOWZ:  # Must have a sim type that starts with IA_ or is called exactly IA. Its how it gets divided into Ia or contaminant.
