+++ examples/global.yml
@@ -1,3 +1,7 @@
+# Note that these examples are for the specific task only, and you cannot run this file (as there are no dependent tasks).
+# Use these files to see how you can configure your tasks, and the configs directory contains full analyses showing
+# all the tasks and how they relate to each other
+
 # You can also overwrite the global settings normally found in the cfg.yml file
 # in your individual runs. Here is an example which reduces the number of jobs
 # allowed in the queues and also adds another data path to look in when searching
