### 2020-03-10 #2122 @temuller
Hi Sam, thanks for the prompt reply.

I know installing SNANA can be a hassle (I have installed it myself!), but as it is a core package used by Pippin, you should be able to provide a straight forward example (assuming the user has SNANA already installed). If you think recording a video can show the usage properly, there is no problem with that, but I am not sure if it is actually simpler that way.
