### 2020-03-10 #2122 @Samreay
Ah, thanks for the catch on the hyperlink I broke! And ah, if you have a working SNANA install, I am hoping that this config file would run without hassle for you: https://github.com/Samreay/Pippin/blob/master/configs/examples/example_sim.yml

If there are issues I'll record the video - Im conscious about not taking too much of your time installing deps and smacking paths around given how much work you've done already. Thanks for the comments so far too, appreciate it!
