### 2020-03-10 #2122 @temuller
Pippin is complaining about the "group" keyword for SNANA in the cfg.yml file. It doesn't seem to find it anywhere. Is it meant to work with midway?
