### 2020-03-10 #2122 @Samreay
Ah, yes, it's trying to set the group ownership over newly created files to Rick Kessler's group by default (the one we use in DES) so that other members of the collaboration can use it. You should be able to set the group to your own name or whatever group you are part of
