### 2020-03-10 #2122 @temuller
It doesn't find any group if I run it locally on my machine
