### 2020-03-10 #2122 @Samreay
As in it can't find the property itself? Throw me the stack if you can,
I'll get it fixed up asap
