### 2020-03-10 #2122 @temuller
Here is the output
[pippin_output.txt](https://github.com/openjournals/joss-reviews/files/4312986/pippin_output.txt)
