### 2020-03-10 #2122 @Samreay
Ah, so yeah we need to find you a valid group - you can list groups in linux using `groups $USER`. The second error is then because Pippin needs python3.7+ as per the doco alas - text is a new kwarg. Let me make a quick video showing it in action so you dont have to dick around.
