### 2020-03-10 #2122 @temuller
Thanks for the video Sam. I was able to correct the errors I had, but I encountered a new one (it has to do with subprocess.check_output). However, I think that after solving this one, the example should run fine. See below.

[pippin_output2.txt](https://github.com/openjournals/joss-reviews/files/4315325/pippin_output2.txt)
