### 2020-03-10 #2122 @Samreay
Ah yes, so Pippin is supposed to be run on an HPC, as it submits and checks
on the status of jobs. To do that it executes squeue, which seems to be
having an issue - are you on an HPC?
