### 2020-03-18 #2122 @kboone
I have attached my review. I have a couple of minor suggestions that should be fixed before accepting this paper for publication.

[JOSS Review #2122.pdf](https://github.com/openjournals/joss-reviews/files/4346420/JOSS.Review.2122.pdf)
