### 2020-03-18 7e4c108
Updating readme to respond to Kyles feedback
