### 2020-03-18 #2122 @Samreay
Hey Kyle! I've made the requested changes to improve the documentation and pushed them out. You can verify the changes here: https://github.com/Samreay/Pippin/commit/7e4c108d086e00c79d436ddc37dd92dc824f0062

Cheers!
