### 2020-03-18 69a39c5
Adding more debug output in case Helen's issue happens again
