### 2020-03-18 #2122 @kboone
The changes look good, and they address my concerns. I recommend this paper for acceptance.
