### 2020-03-19 #2122 @arfon
@Samreay - a few comments on your paper:

Do you want the title to be only 'Pippin'? _Pippin: A pipeline for supernova cosmology_ seems more descriptive?

>  For the DES-3YR analysis (Abbott et al., 2019), we implemented a basic bash pipeline
to reduce manual labour. For the DES-5YR analysis, which greatly increases complexity, I
implement this pipeline to increase researcher productivity.

This ☝️ paragraph reads a little weirdly. You're using 'we' and then 'I' in the same sentence. Perhaps this could be rephrased as:

>  For the DES-3YR analysis (Abbott et al., 2019), we implemented a basic bash pipeline
to reduce manual labour. For the DES-5YR analysis, which greatly increases complexity, Pippin was developed as the pipeline to increase researcher productivity.

Should 'SALT2 supernova model' have a citation?

Last sentence change Github to GitHub. Also, consider linking the phrase 'development page' to to https://github.com/Samreay/Pippin/issues .
