### 2020-03-19 10103c0
Updating Github name and adding external link to documentation
