### 2020-03-19 #2122 @Samreay
Hi @arfon, I've made the I/we consistent change in the latest commit. The SALT2 paper reference is still in draft and yet to be submitted. Github -> GitHub, and a link to the developer doco has been added.
