### 2020-03-19 #2122 @arfon
@Samreay - last time I'll ask I promise (as the title of the Zenodo release is _Pippin: A pipeline for supernova cosmology_. You definitely want the title of the JOSS paper to only be _Pippin_?
