### 2020-03-19 #2122 @Samreay
Oh I thought the title was the full "Pippin: A pipeline for supernova cosmology", as it is in this review. Definitely the full name makes more sense haha
